refactor(semantic): split 3,256-line semantic.rs god file into 6 sub-modules
Split `crates/perl-semantic-analyzer/src/analysis/semantic.rs` (3,256 LOC,
67 functions) into a directory module with focused sub-modules:

- `tokens.rs`      — SemanticTokenType, SemanticTokenModifier, SemanticToken
- `hover.rs`       — HoverInfo struct
- `builtins.rs`    — BuiltinDoc, get_builtin_documentation, classification helpers
- `node_analysis.rs` — impl SemanticAnalyzer: analyze_node (876-line AST match)
                       and all source-text helper methods, infer_type
- `references.rs`  — impl SemanticAnalyzer: resolve_reference_to_symbols,
                     find_all_references, is_symbol_visible
- `model.rs`       — SemanticModel (stable LSP facade)
- `mod.rs`         — SemanticAnalyzer struct (pub(super) fields), core API
                     methods (analyze, symbol_table, semantic_tokens, hover_at,
                     find_definition, is_file_test_operator), module re-exports,
                     all 33 tests

The public API is unchanged — all `pub use` re-exports preserve downstream
consumers. Fields are `pub(super)` so sibling impl blocks can access them.

Closes #2293
