--- crates/perl-semantic-analyzer/src/analysis/semantic.rs
@@ -1,3333 +1,0 @@
-//! Semantic analysis for IDE features
-//!
-//! This module provides semantic analysis on top of the symbol table,
-//! including semantic tokens for syntax highlighting, hover information,
-//! and code intelligence features.
-
-use crate::SourceLocation;
-use crate::ast::{Node, NodeKind};
-use crate::symbol::{ScopeId, ScopeKind, Symbol, SymbolExtractor, SymbolKind, SymbolTable};
-use regex::Regex;
-use std::collections::HashMap;
-use std::sync::OnceLock;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-/// Semantic token types for syntax highlighting in the Parse/Complete workflow.
-pub enum SemanticTokenType {
-    // Variables
-    /// Variable reference (scalar, array, or hash)
-    Variable,
-    /// Variable declaration site
-    VariableDeclaration,
-    /// Read-only variable (constant)
-    VariableReadonly,
-    /// Function parameter
-    Parameter,
-
-    // Functions
-    /// Function/subroutine reference
-    Function,
-    /// Function/subroutine declaration
-    FunctionDeclaration,
-    /// Object method call
-    Method,
-
-    // Types
-    /// Class/package name
-    Class,
-    /// Package namespace
-    Namespace,
-    /// Type annotation (modern Perl)
-    Type,
-
-    // Keywords
-    /// Language keyword (if, while, etc.)
-    Keyword,
-    /// Control flow keyword (return, next, last)
-    KeywordControl,
-    /// Variable modifier (my, our, local, state)
-    Modifier,
-
-    // Literals
-    /// Numeric literal
-    Number,
-    /// String literal
-    String,
-    /// Regular expression
-    Regex,
-
-    // Comments
-    /// Regular comment
-    Comment,
-    /// Documentation comment (POD)
-    CommentDoc,
-
-    // Other
-    /// Operator (+, -, =~, etc.)
-    Operator,
-    /// Punctuation marks and delimiters
-    Punctuation,
-    /// Code label for goto statements
-    Label,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-/// Semantic token modifiers for Analyze/Complete stage highlighting.
-///
-/// Provides additional context about semantic tokens beyond their base type,
-/// enabling rich editor highlighting with detailed symbol information.
-///
-/// # LSP Integration
-/// Maps to LSP `SemanticTokenModifiers` for consistent editor experience
-/// across different LSP clients with full Perl language semantics.
-pub enum SemanticTokenModifier {
-    /// Symbol is being declared at this location
-    Declaration,
-    /// Symbol is being defined at this location
-    Definition,
-    /// Symbol is read-only (constant)
-    Readonly,
-    /// Symbol has static storage duration (state variables)
-    Static,
-    /// Symbol is deprecated and should not be used
-    Deprecated,
-    /// Symbol is abstract (method without implementation)
-    Abstract,
-    /// Symbol represents an asynchronous operation
-    Async,
-    /// Symbol is being modified (written to)
-    Modification,
-    /// Symbol is documentation-related (POD)
-    Documentation,
-    /// Symbol is from the Perl standard library
-    DefaultLibrary,
-}
-
-#[derive(Debug, Clone)]
-/// A semantic token with type and modifiers for LSP syntax highlighting.
-///
-/// Represents a single semantic unit in Perl source code with precise location
-/// and rich type information for enhanced editor experience.
-///
-/// # Performance Characteristics
-/// - Memory: ~32 bytes per token (optimized for large files)
-/// - Serialization: Direct LSP protocol mapping
-/// - Batch processing: Efficient delta updates for incremental parsing
-///
-/// # LSP Workflow Integration
-/// Core component in Parse → Index → Navigate → Complete → Analyze pipeline
-/// for real-time syntax highlighting with ≤1ms update latency.
-pub struct SemanticToken {
-    /// Source location of the token
-    pub location: SourceLocation,
-    /// Semantic classification of the token
-    pub token_type: SemanticTokenType,
-    /// Additional modifiers for enhanced highlighting
-    pub modifiers: Vec<SemanticTokenModifier>,
-}
-
-#[derive(Debug, Clone)]
-/// Hover information for symbols displayed in LSP hover requests.
-///
-/// Provides comprehensive symbol information including signature,
-/// documentation, and contextual details for enhanced developer experience.
-///
-/// Used during Navigate/Analyze stages to answer hover queries.
-///
-/// # Performance Characteristics
-/// - Computation: <100μs for typical symbol lookup
-/// - Memory: Cached per symbol for repeated access
-/// - LSP response: <50ms end-to-end including network
-///
-/// # Perl Context Integration
-/// - Subroutine signatures with parameter information
-/// - Package qualification and scope context
-/// - POD documentation extraction and formatting
-/// - Variable type inference and usage patterns
-///
-/// Workflow: Navigate/Analyze hover details for LSP.
-pub struct HoverInfo {
-    /// Symbol signature or declaration
-    pub signature: String,
-    /// Documentation extracted from POD or comments
-    pub documentation: Option<String>,
-    /// Additional contextual details
-    pub details: Vec<String>,
-}
-
-#[derive(Debug)]
-/// Semantic analyzer providing comprehensive IDE features for Perl code.
-///
-/// Central component for LSP semantic analysis, combining symbol table
-/// construction, semantic token generation, and hover information extraction
-/// with enterprise-grade performance characteristics.
-///
-/// # Performance Characteristics
-/// - Analysis time: O(n) where n is AST node count
-/// - Memory usage: ~1MB per 10K lines of Perl code
-/// - Incremental updates: ≤1ms for typical changes
-/// - Symbol resolution: <50μs average lookup time
-///
-/// # LSP Workflow Integration
-/// Core pipeline component:
-/// 1. **Parse**: AST generation from Perl source
-/// 2. **Index**: Symbol table and semantic token construction
-/// 3. **Navigate**: Symbol resolution for go-to-definition
-/// 4. **Complete**: Context-aware completion suggestions
-/// 5. **Analyze**: Cross-reference analysis and diagnostics
-///
-/// # Perl Language Support
-/// - Full Perl 5 syntax coverage with modern idioms
-/// - Package-qualified symbol resolution
-/// - Lexical scoping with `my`, `our`, `local`, `state`
-/// - Object-oriented method dispatch
-/// - Regular expression and heredoc analysis
-///
-/// # Examples
-///
-/// ```ignore
-/// use perl_parser::{Parser, SemanticAnalyzer};
-///
-/// let code = "my $greeting = 'hello'; sub say_hi { print $greeting; }";
-/// let mut parser = Parser::new(code);
-/// let ast = parser.parse()?;
-///
-/// let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-/// let symbols = analyzer.symbol_table();
-/// let tokens = analyzer.semantic_tokens();
-/// ```
-pub struct SemanticAnalyzer {
-    /// Symbol table with scope hierarchy and definitions
-    symbol_table: SymbolTable,
-    /// Generated semantic tokens for syntax highlighting
-    semantic_tokens: Vec<SemanticToken>,
-    /// Hover information cache for symbol details
-    hover_info: HashMap<SourceLocation, HoverInfo>,
-    /// Source code for text extraction and analysis
-    source: String,
-}
-
-impl SemanticAnalyzer {
-    /// Create a new semantic analyzer from an AST.
-    ///
-    /// Equivalent to [`analyze_with_source`](Self::analyze_with_source) with an empty
-    /// source string. Use this when you only need symbol-table and token analysis
-    /// without source-text-dependent features like hover documentation.
-    ///
-    /// # Examples
-    ///
-    /// ```ignore
-    /// use perl_parser::{Parser, SemanticAnalyzer};
-    ///
-    /// let mut parser = Parser::new("my $x = 42;");
-    /// let ast = parser.parse()?;
-    /// let analyzer = SemanticAnalyzer::analyze(&ast);
-    /// assert!(!analyzer.symbol_table().symbols.is_empty());
-    /// ```
-    pub fn analyze(ast: &Node) -> Self {
-        Self::analyze_with_source(ast, "")
-    }
-
-    /// Create a new semantic analyzer from an AST and source text.
-    ///
-    /// The source text enables richer analysis including hover documentation
-    /// extraction and precise text-range lookups.
-    ///
-    /// # Examples
-    ///
-    /// ```ignore
-    /// use perl_parser::{Parser, SemanticAnalyzer};
-    ///
-    /// let code = "sub greet { print \"Hello\\n\"; }";
-    /// let mut parser = Parser::new(code);
-    /// let ast = parser.parse()?;
-    ///
-    /// let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-    /// let tokens = analyzer.semantic_tokens();
-    /// // tokens contains semantic highlighting data for the parsed code
-    /// ```
-    pub fn analyze_with_source(ast: &Node, source: &str) -> Self {
-        let symbol_table = SymbolExtractor::new_with_source(source).extract(ast);
-
-        let mut analyzer = SemanticAnalyzer {
-            symbol_table,
-            semantic_tokens: Vec::new(),
-            hover_info: HashMap::new(),
-            source: source.to_string(),
-        };
-
-        analyzer.analyze_node(ast, 0);
-        analyzer
-    }
-
-    /// Get the symbol table
-    pub fn symbol_table(&self) -> &SymbolTable {
-        &self.symbol_table
-    }
-
-    /// Get semantic tokens for syntax highlighting
-    pub fn semantic_tokens(&self) -> &[SemanticToken] {
-        &self.semantic_tokens
-    }
-
-    /// Get hover information at a location for Navigate/Analyze stages.
-    pub fn hover_at(&self, location: SourceLocation) -> Option<&HoverInfo> {
-        self.hover_info.get(&location)
-    }
-
-    /// Find the symbol at a given location for Navigate workflows.
-    ///
-    /// Returns the most specific (smallest range) symbol that contains the location.
-    /// This ensures that when hovering inside a subroutine body, we return the
-    /// variable at the cursor rather than the enclosing subroutine.
-    pub fn symbol_at(&self, location: SourceLocation) -> Option<&Symbol> {
-        let mut best: Option<&Symbol> = None;
-        let mut best_span = usize::MAX;
-
-        // Search through all symbols for the most specific one at this location
-        for symbols in self.symbol_table.symbols.values() {
-            for symbol in symbols {
-                if symbol.location.start <= location.start && symbol.location.end >= location.end {
-                    let span = symbol.location.end - symbol.location.start;
-                    if span < best_span {
-                        best = Some(symbol);
-                        best_span = span;
-                    }
-                }
-            }
-        }
-        best
-    }
-
-    /// Find the definition of a symbol at a given position for Navigate workflows.
-    pub fn find_definition(&self, position: usize) -> Option<&Symbol> {
-        // First, find if there's a reference at this position
-        for refs in self.symbol_table.references.values() {
-            for reference in refs {
-                if reference.location.start <= position && reference.location.end >= position {
-                    let symbols = self.resolve_reference_to_symbols(reference);
-                    if let Some(first_symbol) = symbols.first() {
-                        return Some(first_symbol);
-                    }
-                }
-            }
-        }
-
-        // If no reference found, check if we're on a definition itself
-        self.symbol_at(SourceLocation { start: position, end: position })
-    }
-
-    /// Resolve a reference to its symbol definitions, handling cross-package lookups
-    fn resolve_reference_to_symbols(
-        &self,
-        reference: &crate::symbol::SymbolReference,
-    ) -> Vec<&Symbol> {
-        // Handle qualified names like Foo::bar
-        if let Some((pkg, name)) = reference.name.rsplit_once("::") {
-            if let Some(pkg_syms) = self.symbol_table.symbols.get(pkg) {
-                let mut results = Vec::new();
-                for sym in pkg_syms {
-                    if sym.kind == SymbolKind::Package {
-                        // Find the scope associated with this package symbol
-                        let pkg_scope = self
-                            .symbol_table
-                            .scopes
-                            .values()
-                            .find(|s| {
-                                s.kind == ScopeKind::Package
-                                    && s.location.start == sym.location.start
-                                    && s.location.end == sym.location.end
-                            })
-                            .map(|s| s.id)
-                            .unwrap_or(sym.scope_id);
-                        // Symbols may live in an inner block scope
-                        let search_scope = self
-                            .symbol_table
-                            .scopes
-                            .values()
-                            .find(|s| s.parent == Some(pkg_scope))
-                            .map(|s| s.id)
-                            .unwrap_or(pkg_scope);
-                        results.extend(self.symbol_table.find_symbol(
-                            name,
-                            search_scope,
-                            reference.kind,
-                        ));
-                    }
-                }
-                results
-            } else {
-                self.symbol_table.find_symbol(name, reference.scope_id, reference.kind)
-            }
-        } else {
-            self.symbol_table.find_symbol(&reference.name, reference.scope_id, reference.kind)
-        }
-    }
-
-    /// Find all references to a symbol at a given position for Navigate/Analyze workflows.
-    pub fn find_all_references(
-        &self,
-        position: usize,
-        include_declaration: bool,
-    ) -> Vec<SourceLocation> {
-        // First find the symbol at this position (either definition or reference)
-        let symbol = if let Some(def) = self.find_definition(position) {
-            Some(def)
-        } else {
-            // Check if we're on a reference
-            for refs in self.symbol_table.references.values() {
-                for reference in refs {
-                    if reference.location.start <= position && reference.location.end >= position {
-                        // Found a reference, get its definition to get the symbol ID
-                        let symbols = self.symbol_table.find_symbol(
-                            &reference.name,
-                            reference.scope_id,
-                            reference.kind,
-                        );
-                        if let Some(first_symbol) = symbols.first() {
-                            return self
-                                .find_all_references_for_symbol(first_symbol, include_declaration);
-                        }
-                    }
-                }
-            }
-            None
-        };
-
-        if let Some(symbol) = symbol {
-            return self.find_all_references_for_symbol(symbol, include_declaration);
-        }
-
-        Vec::new()
-    }
-
-    /// Find all references for a specific symbol
-    fn find_all_references_for_symbol(
-        &self,
-        symbol: &Symbol,
-        include_declaration: bool,
-    ) -> Vec<SourceLocation> {
-        let mut locations = Vec::new();
-
-        // Include the declaration if requested
-        if include_declaration {
-            locations.push(symbol.location);
-        }
-
-        // Find all references to this symbol by name
-        if let Some(refs) = self.symbol_table.references.get(&symbol.name) {
-            for reference in refs {
-                // Only include references of the same kind and in scope where the symbol is visible
-                if reference.kind == symbol.kind {
-                    // Check if the symbol is visible from this reference's scope
-                    if self.is_symbol_visible(symbol, reference.scope_id) {
-                        locations.push(reference.location);
-                    }
-                }
-            }
-        }
-
-        locations
-    }
-
-    /// Check if a symbol is visible from a given scope
-    fn is_symbol_visible(&self, symbol: &Symbol, scope_id: ScopeId) -> bool {
-        // For now, simple visibility check:
-        // - Symbols in the same scope are visible
-        // - Symbols in parent scopes are visible
-        // - Package-level symbols are visible from package scopes
-
-        if symbol.scope_id == scope_id {
-            return true;
-        }
-
-        // Check if scope_id is a descendant of symbol.scope_id
-        let mut current_scope = scope_id;
-        while let Some(scope) = self.symbol_table.scopes.get(&current_scope) {
-            if scope.parent == Some(symbol.scope_id) {
-                return true;
-            }
-            if let Some(parent) = scope.parent {
-                current_scope = parent;
-            } else {
-                break;
-            }
-        }
-
-        // For package-level symbols (scope_id 0), always visible
-        symbol.scope_id == 0
-    }
-
-    /// Check if an operator is a file test operator.
-    ///
-    /// File test operators in Perl are unary operators that test file properties:
-    /// -e (exists), -d (directory), -f (file), -r (readable), -w (writable), etc.
-    fn is_file_test_operator(op: &str) -> bool {
-        matches!(
-            op,
-            "-e" | "-d"
-                | "-f"
-                | "-r"
-                | "-w"
-                | "-x"
-                | "-s"
-                | "-z"
-                | "-T"
-                | "-B"
-                | "-M"
-                | "-A"
-                | "-C"
-                | "-l"
-                | "-p"
-                | "-S"
-                | "-u"
-                | "-g"
-                | "-k"
-                | "-t"
-                | "-O"
-                | "-G"
-                | "-R"
-                | "-b"
-                | "-c"
-        )
-    }
-
-    /// Analyze a node and generate semantic information
-    fn analyze_node(&mut self, node: &Node, scope_id: ScopeId) {
-        match &node.kind {
-            NodeKind::Program { statements } => {
-                for stmt in statements {
-                    self.analyze_node(stmt, scope_id);
-                }
-            }
-
-            NodeKind::VariableDeclaration { declarator, variable, attributes, initializer } => {
-                // Add semantic token for declaration
-                if let NodeKind::Variable { sigil, name } = &variable.kind {
-                    let token_type = match declarator.as_str() {
-                        "my" | "state" => SemanticTokenType::VariableDeclaration,
-                        "our" => SemanticTokenType::Variable,
-                        "local" => SemanticTokenType::Variable,
-                        _ => SemanticTokenType::Variable,
-                    };
-
-                    let mut modifiers = vec![SemanticTokenModifier::Declaration];
-                    if declarator == "state" || attributes.iter().any(|a| a == ":shared") {
-                        modifiers.push(SemanticTokenModifier::Static);
-                    }
-
-                    self.semantic_tokens.push(SemanticToken {
-                        location: variable.location,
-                        token_type,
-                        modifiers,
-                    });
-
-                    // Add hover info
-                    let hover = HoverInfo {
-                        signature: format!("{} {}{}", declarator, sigil, name),
-                        documentation: self.extract_documentation(node.location.start),
-                        details: format_attribute_details(attributes),
-                    };
-
-                    self.hover_info.insert(variable.location, hover);
-                }
-
-                if let Some(init) = initializer {
-                    self.analyze_node(init, scope_id);
-                }
-            }
-
-            NodeKind::Variable { sigil, name } => {
-                let kind = match sigil.as_str() {
-                    "$" => SymbolKind::scalar(),
-                    "@" => SymbolKind::array(),
-                    "%" => SymbolKind::hash(),
-                    _ => return,
-                };
-
-                // Find the symbol definition
-                let symbols = self.symbol_table.find_symbol(name, scope_id, kind);
-
-                let token_type = if let Some(symbol) = symbols.first() {
-                    match symbol.declaration.as_deref() {
-                        Some("my") | Some("state") => SemanticTokenType::Variable,
-                        Some("our") => SemanticTokenType::Variable,
-                        _ => SemanticTokenType::Variable,
-                    }
-                } else {
-                    // Undefined variable
-                    SemanticTokenType::Variable
-                };
-
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type,
-                    modifiers: vec![],
-                });
-
-                // Add hover info if we found the symbol
-                if let Some(symbol) = symbols.first() {
-                    let hover = HoverInfo {
-                        signature: format!(
-                            "{} {}{}",
-                            symbol.declaration.as_deref().unwrap_or(""),
-                            sigil,
-                            name
-                        )
-                        .trim()
-                        .to_string(),
-                        documentation: symbol.documentation.clone(),
-                        details: vec![format!(
-                            "Defined at line {}",
-                            self.line_number(symbol.location.start)
-                        )],
-                    };
-
-                    self.hover_info.insert(node.location, hover);
-                }
-            }
-
-            NodeKind::Subroutine { name, prototype, signature, attributes, body, name_span: _ } => {
-                if let Some(sub_name) = name {
-                    // Named subroutine
-                    let token = SemanticToken {
-                        location: node.location,
-                        token_type: SemanticTokenType::FunctionDeclaration,
-                        modifiers: vec![SemanticTokenModifier::Declaration],
-                    };
-
-                    self.semantic_tokens.push(token);
-
-                    // Add hover info
-                    let mut signature_str = format!("sub {}", sub_name);
-                    if signature.is_some() {
-                        signature_str.push_str("(...)");
-                    }
-
-                    let hover = HoverInfo {
-                        signature: signature_str,
-                        documentation: self.extract_documentation(node.location.start),
-                        details: format_attribute_details(attributes),
-                    };
-
-                    self.hover_info.insert(node.location, hover);
-                } else {
-                    // Anonymous subroutine (closure)
-                    // Add semantic token for the 'sub' keyword
-                    self.semantic_tokens.push(SemanticToken {
-                        location: SourceLocation {
-                            start: node.location.start,
-                            end: node.location.start + 3, // "sub"
-                        },
-                        token_type: SemanticTokenType::Keyword,
-                        modifiers: vec![],
-                    });
-
-                    // Add hover info for anonymous subs
-                    let mut signature_str = "sub".to_string();
-                    if signature.is_some() {
-                        signature_str.push_str(" (...)");
-                    }
-                    signature_str.push_str(" { ... }");
-
-                    let mut details = vec!["Anonymous subroutine (closure)".to_string()];
-                    details.extend(format_attribute_details(attributes));
-
-                    let hover = HoverInfo {
-                        signature: signature_str,
-                        documentation: self.extract_documentation(node.location.start),
-                        details,
-                    };
-
-                    self.hover_info.insert(node.location, hover);
-                }
-
-                {
-                    // Get the subroutine scope from the symbol table
-                    let sub_scope = self.get_scope_for(node, ScopeKind::Subroutine);
-
-                    if let Some(proto) = prototype {
-                        self.analyze_node(proto, sub_scope);
-                    }
-                    if let Some(sig) = signature {
-                        self.analyze_node(sig, sub_scope);
-                    }
-
-                    self.analyze_node(body, sub_scope);
-                }
-            }
-
-            NodeKind::Method { name, signature, attributes, body } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location, // Approximate, ideally name span
-                    token_type: SemanticTokenType::FunctionDeclaration,
-                    modifiers: vec![SemanticTokenModifier::Declaration],
-                });
-
-                // Add hover info
-                let hover = HoverInfo {
-                    signature: format!("method {}", name),
-                    documentation: self.extract_documentation(node.location.start),
-                    details: format_attribute_details(attributes),
-                };
-                self.hover_info.insert(node.location, hover);
-
-                // Analyze body in new scope (assumed same as Subroutine scope kind for now)
-                let sub_scope = self.get_scope_for(node, ScopeKind::Subroutine);
-                if let Some(sig) = signature {
-                    self.analyze_node(sig, sub_scope);
-                }
-                self.analyze_node(body, sub_scope);
-            }
-
-            NodeKind::FunctionCall { name, args } => {
-                // Check if this is a built-in function
-                {
-                    let token_type = if is_control_keyword(name) {
-                        SemanticTokenType::KeywordControl
-                    } else if is_builtin_function(name) {
-                        SemanticTokenType::Function
-                    } else {
-                        // Check if it's a user-defined function
-                        let symbols =
-                            self.symbol_table.find_symbol(name, scope_id, SymbolKind::Subroutine);
-                        if symbols.is_empty() {
-                            SemanticTokenType::Function
-                        } else {
-                            SemanticTokenType::Function
-                        }
-                    };
-
-                    self.semantic_tokens.push(SemanticToken {
-                        location: node.location,
-                        token_type,
-                        modifiers: if is_builtin_function(name) && !is_control_keyword(name) {
-                            vec![SemanticTokenModifier::DefaultLibrary]
-                        } else {
-                            vec![]
-                        },
-                    });
-
-                    // Add hover for built-ins
-                    if let Some(doc) = get_builtin_documentation(name) {
-                        let hover = HoverInfo {
-                            signature: doc.signature.to_string(),
-                            documentation: Some(doc.description.to_string()),
-                            details: vec![],
-                        };
-
-                        self.hover_info.insert(node.location, hover);
-                    }
-                }
-
-                // Name is already a string, not a node
-                for arg in args {
-                    self.analyze_node(arg, scope_id);
-                }
-            }
-
-            NodeKind::Package { name, block, name_span: _ } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::Namespace,
-                    modifiers: vec![SemanticTokenModifier::Declaration],
-                });
-
-                // Try POD docs first, then fall back to leading comments
-                let documentation = self
-                    .extract_pod_name_section(name)
-                    .or_else(|| self.extract_documentation(node.location.start));
-
-                let hover = HoverInfo {
-                    signature: format!("package {}", name),
-                    documentation,
-                    details: vec![],
-                };
-
-                self.hover_info.insert(node.location, hover);
-
-                if let Some(block_node) = block {
-                    let package_scope = self.get_scope_for(node, ScopeKind::Package);
-                    self.analyze_node(block_node, package_scope);
-                }
-            }
-
-            NodeKind::String { value: _, interpolated: _ } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::String,
-                    modifiers: vec![],
-                });
-            }
-
-            NodeKind::Number { value: _ } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::Number,
-                    modifiers: vec![],
-                });
-            }
-
-            NodeKind::Regex { .. } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::Regex,
-                    modifiers: vec![],
-                });
-            }
-
-            NodeKind::Match { expr, .. } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::Regex,
-                    modifiers: vec![],
-                });
-                self.analyze_node(expr, scope_id);
-            }
-            NodeKind::Substitution { expr, .. } => {
-                // Substitution operator: s/// - add semantic token for the operator
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::Operator,
-                    modifiers: vec![],
-                });
-                self.analyze_node(expr, scope_id);
-            }
-            NodeKind::Transliteration { expr, .. } => {
-                // Transliteration operator: tr/// or y/// - add semantic token for the operator
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::Operator,
-                    modifiers: vec![],
-                });
-                self.analyze_node(expr, scope_id);
-            }
-
-            NodeKind::LabeledStatement { label: _, statement } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::Label,
-                    modifiers: vec![],
-                });
-
-                {
-                    self.analyze_node(statement, scope_id);
-                }
-            }
-
-            // Control flow keywords
-            NodeKind::If { condition, then_branch, elsif_branches, else_branch } => {
-                self.analyze_node(condition, scope_id);
-                self.analyze_node(then_branch, scope_id);
-                for (elsif_cond, elsif_branch) in elsif_branches {
-                    self.analyze_node(elsif_cond, scope_id);
-                    self.analyze_node(elsif_branch, scope_id);
-                }
-                if let Some(else_node) = else_branch {
-                    self.analyze_node(else_node, scope_id);
-                }
-            }
-
-            NodeKind::While { condition, body, continue_block: _ } => {
-                self.analyze_node(condition, scope_id);
-                self.analyze_node(body, scope_id);
-            }
-
-            NodeKind::For { init, condition, update, body, .. } => {
-                if let Some(init_node) = init {
-                    self.analyze_node(init_node, scope_id);
-                }
-                if let Some(cond_node) = condition {
-                    self.analyze_node(cond_node, scope_id);
-                }
-                if let Some(update_node) = update {
-                    self.analyze_node(update_node, scope_id);
-                }
-                self.analyze_node(body, scope_id);
-            }
-
-            NodeKind::Foreach { variable, list, body, continue_block } => {
-                self.analyze_node(variable, scope_id);
-                self.analyze_node(list, scope_id);
-                self.analyze_node(body, scope_id);
-                if let Some(cb) = continue_block {
-                    self.analyze_node(cb, scope_id);
-                }
-            }
-
-            // Recursively analyze other nodes
-            NodeKind::Block { statements } => {
-                for stmt in statements {
-                    self.analyze_node(stmt, scope_id);
-                }
-            }
-
-            NodeKind::Binary { left, right, .. } => {
-                self.analyze_node(left, scope_id);
-                self.analyze_node(right, scope_id);
-            }
-
-            NodeKind::Assignment { lhs, rhs, .. } => {
-                self.analyze_node(lhs, scope_id);
-                self.analyze_node(rhs, scope_id);
-            }
-
-            // Phase 1: Critical LSP Features (Issue #188)
-            NodeKind::VariableListDeclaration {
-                declarator,
-                variables,
-                attributes,
-                initializer,
-            } => {
-                // Handle multi-variable declarations like: my ($x, $y, $z) = (1, 2, 3);
-                for var in variables {
-                    if let NodeKind::Variable { sigil, name } = &var.kind {
-                        let token_type = match declarator.as_str() {
-                            "my" | "state" => SemanticTokenType::VariableDeclaration,
-                            "our" => SemanticTokenType::Variable,
-                            "local" => SemanticTokenType::Variable,
-                            _ => SemanticTokenType::Variable,
-                        };
-
-                        let mut modifiers = vec![SemanticTokenModifier::Declaration];
-                        if declarator == "state" || attributes.iter().any(|a| a == ":shared") {
-                            modifiers.push(SemanticTokenModifier::Static);
-                        }
-
-                        self.semantic_tokens.push(SemanticToken {
-                            location: var.location,
-                            token_type,
-                            modifiers,
-                        });
-
-                        // Add hover info
-                        let hover = HoverInfo {
-                            signature: format!("{} {}{}", declarator, sigil, name),
-                            documentation: self.extract_documentation(var.location.start),
-                            details: format_attribute_details(attributes),
-                        };
-
-                        self.hover_info.insert(var.location, hover);
-                    }
-                }
-
-                if let Some(init) = initializer {
-                    self.analyze_node(init, scope_id);
-                }
-            }
-
-            NodeKind::Ternary { condition, then_expr, else_expr } => {
-                // Handle conditional expressions: $x ? $y : $z
-                self.analyze_node(condition, scope_id);
-                self.analyze_node(then_expr, scope_id);
-                self.analyze_node(else_expr, scope_id);
-            }
-
-            NodeKind::ArrayLiteral { elements } => {
-                // Handle array constructors: [1, 2, 3, 4]
-                for elem in elements {
-                    self.analyze_node(elem, scope_id);
-                }
-            }
-
-            NodeKind::HashLiteral { pairs } => {
-                // Handle hash constructors: { key1 => "value1", key2 => "value2" }
-                for (key, value) in pairs {
-                    self.analyze_node(key, scope_id);
-                    self.analyze_node(value, scope_id);
-                }
-            }
-
-            NodeKind::Try { body, catch_blocks, finally_block } => {
-                // Handle try/catch error handling
-                self.analyze_node(body, scope_id);
-
-                for (_var, catch_body) in catch_blocks {
-                    // Note: var is just a String (variable name), not a Node
-                    self.analyze_node(catch_body, scope_id);
-                }
-
-                if let Some(finally) = finally_block {
-                    self.analyze_node(finally, scope_id);
-                }
-            }
-
-            NodeKind::PhaseBlock { phase: _, phase_span: _, block } => {
-                // Handle BEGIN/END/INIT/CHECK/UNITCHECK blocks
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::Keyword,
-                    modifiers: vec![],
-                });
-
-                self.analyze_node(block, scope_id);
-            }
-
-            NodeKind::ExpressionStatement { expression } => {
-                // Handle expression statements: $x + 10;
-                // Just delegate to the wrapped expression
-                self.analyze_node(expression, scope_id);
-            }
-
-            NodeKind::Do { block } => {
-                // Handle do blocks: do { ... }
-                // Do blocks create expression context but maintain scope
-                self.analyze_node(block, scope_id);
-            }
-
-            NodeKind::Eval { block } => {
-                // Handle eval blocks: eval { dangerous_operation(); }
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::Keyword,
-                    modifiers: vec![],
-                });
-
-                // Eval blocks should create a new scope for error isolation
-                self.analyze_node(block, scope_id);
-            }
-
-            NodeKind::VariableWithAttributes { variable, attributes } => {
-                // Handle attributed variables: my $x :shared = 42;
-                // Analyze the base variable node
-                self.analyze_node(variable, scope_id);
-
-                // Add modifier tokens for special attributes
-                if attributes.iter().any(|a| a == ":shared" || a == ":lvalue") {
-                    // The variable node was already processed, so we just note the attributes
-                    // in the hover info (if we need to enhance it later)
-                }
-            }
-
-            NodeKind::Unary { op, operand } => {
-                // Handle unary operators: -$x, !$x, ++$x, $x++
-                // Add token for the operator itself (if needed for highlighting)
-                if matches!(op.as_str(), "++" | "--" | "!" | "-" | "~" | "\\") {
-                    self.semantic_tokens.push(SemanticToken {
-                        location: node.location,
-                        token_type: SemanticTokenType::Operator,
-                        modifiers: vec![],
-                    });
-                }
-
-                // Handle file test operators: -e, -d, -f, -r, -w, -x, -s, -z, -T, -B, etc.
-                if Self::is_file_test_operator(op) {
-                    self.semantic_tokens.push(SemanticToken {
-                        location: node.location,
-                        token_type: SemanticTokenType::Operator,
-                        modifiers: vec![],
-                    });
-                }
-
-                self.analyze_node(operand, scope_id);
-            }
-
-            NodeKind::Readline { filehandle } => {
-                // Handle readline/diamond operator: <STDIN>, <$fh>, <>
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::Operator, // diamond operator is an I/O operator
-                    modifiers: vec![],
-                });
-
-                // Add hover info for common filehandles
-                if let Some(fh) = filehandle {
-                    let hover = HoverInfo {
-                        signature: format!("<{}>", fh),
-                        documentation: match fh.as_str() {
-                            "STDIN" => Some("Standard input filehandle".to_string()),
-                            "STDOUT" => Some("Standard output filehandle".to_string()),
-                            "STDERR" => Some("Standard error filehandle".to_string()),
-                            _ => Some(format!("Read from filehandle {}", fh)),
-                        },
-                        details: vec![],
-                    };
-                    self.hover_info.insert(node.location, hover);
-                } else {
-                    // Bare <> reads from ARGV or STDIN
-                    let hover = HoverInfo {
-                        signature: "<>".to_string(),
-                        documentation: Some("Read from command-line files or STDIN".to_string()),
-                        details: vec![],
-                    };
-                    self.hover_info.insert(node.location, hover);
-                }
-            }
-
-            // Phase 2/3 Handlers
-            NodeKind::MethodCall { object, method, args } => {
-                self.analyze_node(object, scope_id);
-
-                if let Some(offset) =
-                    self.find_substring_in_source_after(node, method, object.location.end)
-                {
-                    self.semantic_tokens.push(SemanticToken {
-                        location: SourceLocation { start: offset, end: offset + method.len() },
-                        token_type: SemanticTokenType::Method,
-                        modifiers: vec![],
-                    });
-                }
-
-                for arg in args {
-                    self.analyze_node(arg, scope_id);
-                }
-            }
-
-            NodeKind::IndirectCall { method, object, args } => {
-                if let Some(offset) = self.find_method_name_in_source(node, method) {
-                    self.semantic_tokens.push(SemanticToken {
-                        location: SourceLocation { start: offset, end: offset + method.len() },
-                        token_type: SemanticTokenType::Method,
-                        modifiers: vec![],
-                    });
-                }
-                self.analyze_node(object, scope_id);
-                for arg in args {
-                    self.analyze_node(arg, scope_id);
-                }
-            }
-
-            NodeKind::Use { module, args, .. } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: SourceLocation {
-                        start: node.location.start,
-                        end: node.location.start + 3,
-                    },
-                    token_type: SemanticTokenType::Keyword,
-                    modifiers: vec![],
-                });
-
-                let mut args_start = node.location.start + 3;
-                if let Some(offset) = self.find_substring_in_source(node, module) {
-                    self.semantic_tokens.push(SemanticToken {
-                        location: SourceLocation { start: offset, end: offset + module.len() },
-                        token_type: SemanticTokenType::Namespace,
-                        modifiers: vec![],
-                    });
-                    args_start = offset + module.len();
-                }
-
-                self.analyze_string_args(node, args, args_start);
-            }
-
-            NodeKind::No { module, args, .. } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: SourceLocation {
-                        start: node.location.start,
-                        end: node.location.start + 2,
-                    },
-                    token_type: SemanticTokenType::Keyword,
-                    modifiers: vec![],
-                });
-
-                let mut args_start = node.location.start + 2;
-                if let Some(offset) = self.find_substring_in_source(node, module) {
-                    self.semantic_tokens.push(SemanticToken {
-                        location: SourceLocation { start: offset, end: offset + module.len() },
-                        token_type: SemanticTokenType::Namespace,
-                        modifiers: vec![],
-                    });
-                    args_start = offset + module.len();
-                }
-
-                self.analyze_string_args(node, args, args_start);
-            }
-
-            NodeKind::Given { expr, body } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: SourceLocation {
-                        start: node.location.start,
-                        end: node.location.start + 5,
-                    }, // given
-                    token_type: SemanticTokenType::KeywordControl,
-                    modifiers: vec![],
-                });
-                self.analyze_node(expr, scope_id);
-                self.analyze_node(body, scope_id);
-            }
-
-            NodeKind::When { condition, body } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: SourceLocation {
-                        start: node.location.start,
-                        end: node.location.start + 4,
-                    }, // when
-                    token_type: SemanticTokenType::KeywordControl,
-                    modifiers: vec![],
-                });
-                self.analyze_node(condition, scope_id);
-                self.analyze_node(body, scope_id);
-            }
-
-            NodeKind::Default { body } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: SourceLocation {
-                        start: node.location.start,
-                        end: node.location.start + 7,
-                    }, // default
-                    token_type: SemanticTokenType::KeywordControl,
-                    modifiers: vec![],
-                });
-                self.analyze_node(body, scope_id);
-            }
-
-            NodeKind::Return { value } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: SourceLocation {
-                        start: node.location.start,
-                        end: node.location.start + 6,
-                    }, // return
-                    token_type: SemanticTokenType::KeywordControl,
-                    modifiers: vec![],
-                });
-                if let Some(v) = value {
-                    self.analyze_node(v, scope_id);
-                }
-            }
-
-            NodeKind::Class { name, body } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: SourceLocation {
-                        start: node.location.start,
-                        end: node.location.start + 5,
-                    }, // class
-                    token_type: SemanticTokenType::Keyword,
-                    modifiers: vec![],
-                });
-
-                if let Some(offset) = self.find_substring_in_source(node, name) {
-                    self.semantic_tokens.push(SemanticToken {
-                        location: SourceLocation { start: offset, end: offset + name.len() },
-                        token_type: SemanticTokenType::Class,
-                        modifiers: vec![SemanticTokenModifier::Declaration],
-                    });
-                }
-
-                let class_scope = self.get_scope_for(node, ScopeKind::Package);
-                self.analyze_node(body, class_scope);
-            }
-
-            NodeKind::Signature { parameters } => {
-                for param in parameters {
-                    self.analyze_node(param, scope_id);
-                }
-            }
-
-            NodeKind::MandatoryParameter { variable }
-            | NodeKind::OptionalParameter { variable, .. }
-            | NodeKind::SlurpyParameter { variable }
-            | NodeKind::NamedParameter { variable } => {
-                self.analyze_node(variable, scope_id);
-            }
-
-            NodeKind::Diamond | NodeKind::Ellipsis => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::Operator,
-                    modifiers: vec![],
-                });
-            }
-
-            NodeKind::Undef => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::Keyword,
-                    modifiers: vec![],
-                });
-            }
-
-            NodeKind::Identifier { .. } => {
-                // Bareword identifiers, usually left to lexical highlighting
-                // but we handle them to avoid the default case.
-            }
-
-            NodeKind::Heredoc { .. } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::String,
-                    modifiers: vec![],
-                });
-            }
-
-            NodeKind::Glob { .. } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::Operator,
-                    modifiers: vec![],
-                });
-            }
-
-            NodeKind::DataSection { .. } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::Comment,
-                    modifiers: vec![],
-                });
-            }
-
-            NodeKind::Prototype { .. } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::Punctuation,
-                    modifiers: vec![],
-                });
-            }
-
-            NodeKind::Typeglob { .. } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::Variable,
-                    modifiers: vec![],
-                });
-            }
-
-            NodeKind::Untie { variable } => {
-                self.analyze_node(variable, scope_id);
-            }
-
-            NodeKind::LoopControl { .. } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::KeywordControl,
-                    modifiers: vec![],
-                });
-            }
-
-            NodeKind::Goto { target } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::KeywordControl,
-                    modifiers: vec![],
-                });
-                self.analyze_node(target, scope_id);
-            }
-
-            NodeKind::MissingExpression
-            | NodeKind::MissingStatement
-            | NodeKind::MissingIdentifier
-            | NodeKind::MissingBlock => {
-                // No tokens for missing constructs
-            }
-
-            NodeKind::Tie { variable, package, args } => {
-                self.analyze_node(variable, scope_id);
-                self.analyze_node(package, scope_id);
-                for arg in args {
-                    self.analyze_node(arg, scope_id);
-                }
-            }
-
-            NodeKind::StatementModifier { statement, condition, modifier } => {
-                // Handle postfix loop modifiers: for, while, until, foreach
-                // e.g., print $_ for @list; or $x++ while $x < 10;
-                if matches!(modifier.as_str(), "for" | "foreach" | "while" | "until") {
-                    self.semantic_tokens.push(SemanticToken {
-                        location: node.location,
-                        token_type: SemanticTokenType::KeywordControl,
-                        modifiers: vec![],
-                    });
-                }
-                self.analyze_node(statement, scope_id);
-                self.analyze_node(condition, scope_id);
-            }
-
-            NodeKind::Format { name, .. } => {
-                self.semantic_tokens.push(SemanticToken {
-                    location: node.location,
-                    token_type: SemanticTokenType::FunctionDeclaration,
-                    modifiers: vec![SemanticTokenModifier::Declaration],
-                });
-
-                let hover = HoverInfo {
-                    signature: format!("format {} =", name),
-                    documentation: None,
-                    details: vec![],
-                };
-                self.hover_info.insert(node.location, hover);
-            }
-
-            NodeKind::Error { .. } | NodeKind::UnknownRest => {
-                // No semantic tokens for error nodes
-            }
-        }
-    }
-
-    /// Extract documentation (POD or comments) preceding a position
-    fn extract_documentation(&self, start: usize) -> Option<String> {
-        static POD_RE: OnceLock<Result<Regex, regex::Error>> = OnceLock::new();
-        static COMMENT_RE: OnceLock<Result<Regex, regex::Error>> = OnceLock::new();
-
-        if self.source.is_empty() {
-            return None;
-        }
-        let before = &self.source[..start];
-
-        // Check for POD blocks ending with =cut
-        let pod_re = POD_RE
-            .get_or_init(|| Regex::new(r"(?ms)(=[a-zA-Z0-9].*?\n=cut\n?)\s*$"))
-            .as_ref()
-            .ok()?;
-        if let Some(caps) = pod_re.captures(before) {
-            if let Some(pod_text) = caps.get(1) {
-                return Some(pod_text.as_str().trim().to_string());
-            }
-        }
-
-        // Check for consecutive comment lines
-        let comment_re =
-            COMMENT_RE.get_or_init(|| Regex::new(r"(?m)(#.*\n)+\s*$")).as_ref().ok()?;
-        if let Some(caps) = comment_re.captures(before) {
-            if let Some(comment_match) = caps.get(0) {
-                // Strip the # prefix from each comment line
-                let doc = comment_match
-                    .as_str()
-                    .lines()
-                    .map(|line| line.trim_start_matches('#').trim())
-                    .filter(|line| !line.is_empty())
-                    .collect::<Vec<_>>()
-                    .join(" ");
-                return Some(doc);
-            }
-        }
-
-        None
-    }
-
-    /// Extract the POD `=head1 NAME` section for a package.
-    ///
-    /// Scans the entire source for a `=head1 NAME` POD section and returns
-    /// its content if it mentions the given package name.
-    fn extract_pod_name_section(&self, package_name: &str) -> Option<String> {
-        if self.source.is_empty() {
-            return None;
-        }
-
-        let mut in_name_section = false;
-        let mut name_lines: Vec<&str> = Vec::new();
-
-        for line in self.source.lines() {
-            let trimmed = line.trim();
-            if trimmed.starts_with("=head1") {
-                if in_name_section {
-                    break;
-                }
-                let heading = trimmed.strip_prefix("=head1").map(|s| s.trim());
-                if heading == Some("NAME") {
-                    in_name_section = true;
-                    continue;
-                }
-            } else if trimmed.starts_with("=cut") && in_name_section {
-                break;
-            } else if trimmed.starts_with('=') && in_name_section {
-                break;
-            } else if in_name_section && !trimmed.is_empty() {
-                name_lines.push(trimmed);
-            }
-        }
-
-        if !name_lines.is_empty() {
-            let name_doc = name_lines.join(" ");
-            if name_doc.contains(package_name)
-                || name_doc.contains(&package_name.replace("::", "-"))
-            {
-                return Some(name_doc);
-            }
-        }
-
-        None
-    }
-
-    /// Get scope id for a node by consulting the symbol table
-    fn get_scope_for(&self, node: &Node, kind: ScopeKind) -> ScopeId {
-        for scope in self.symbol_table.scopes.values() {
-            if scope.kind == kind
-                && scope.location.start == node.location.start
-                && scope.location.end == node.location.end
-            {
-                return scope.id;
-            }
-        }
-        0
-    }
-
-    /// Get line number from byte offset (simplified version)
-    fn line_number(&self, offset: usize) -> usize {
-        if self.source.is_empty() { 1 } else { self.source[..offset].lines().count() + 1 }
-    }
-
-    /// Find substring in source within node's range
-    fn find_substring_in_source(&self, node: &Node, substring: &str) -> Option<usize> {
-        if self.source.len() < node.location.end {
-            return None;
-        }
-        let node_text = &self.source[node.location.start..node.location.end];
-        if let Some(pos) = node_text.find(substring) {
-            return Some(node.location.start + pos);
-        }
-        None
-    }
-
-    /// Find method name in source within node's range
-    fn find_method_name_in_source(&self, node: &Node, method_name: &str) -> Option<usize> {
-        self.find_substring_in_source(node, method_name)
-    }
-
-    /// Find substring in source within node's range, starting search after a specific absolute offset
-    fn find_substring_in_source_after(
-        &self,
-        node: &Node,
-        substring: &str,
-        after: usize,
-    ) -> Option<usize> {
-        if self.source.len() < node.location.end || after >= node.location.end {
-            return None;
-        }
-
-        let start_rel = after.saturating_sub(node.location.start);
-
-        let node_text = &self.source[node.location.start..node.location.end];
-        if start_rel >= node_text.len() {
-            return None;
-        }
-
-        let text_to_search = &node_text[start_rel..];
-        if let Some(pos) = text_to_search.find(substring) {
-            return Some(node.location.start + start_rel + pos);
-        }
-        None
-    }
-
-    /// Analyze string arguments for highlighting (e.g. in use/no statements)
-    fn analyze_string_args(&mut self, node: &Node, args: &[String], start_offset: usize) {
-        let mut current_offset = start_offset;
-        for arg in args {
-            if let Some(offset) = self.find_substring_in_source_after(node, arg, current_offset) {
-                self.semantic_tokens.push(SemanticToken {
-                    location: SourceLocation { start: offset, end: offset + arg.len() },
-                    token_type: SemanticTokenType::String,
-                    modifiers: vec![],
-                });
-                current_offset = offset + arg.len();
-            }
-        }
-    }
-
-    /// Infer the type of a node based on its context and initialization
-    ///
-    /// Provides basic type inference for Perl expressions to enhance hover
-    /// information with derived type information. Supports common patterns:
-    /// - Literal values (numbers, strings, arrays, hashes)
-    /// - Variable references (looks up declaration)
-    /// - Function calls (basic return type hints)
-    ///
-    /// In the semantic workflow (Parse -> Index -> Analyze), this method runs
-    /// during the Analyze stage and consumes symbols produced during Index.
-    ///
-    /// # Arguments
-    ///
-    /// * `node` - The AST node to infer type for
-    ///
-    /// # Returns
-    ///
-    /// A string describing the inferred type, or None if type cannot be determined
-    pub fn infer_type(&self, node: &Node) -> Option<String> {
-        match &node.kind {
-            NodeKind::Number { .. } => Some("number".to_string()),
-            NodeKind::String { .. } => Some("string".to_string()),
-            NodeKind::ArrayLiteral { .. } => Some("array".to_string()),
-            NodeKind::HashLiteral { .. } => Some("hash".to_string()),
-
-            NodeKind::Variable { sigil, name } => {
-                // Look up the variable in the symbol table
-                let kind = match sigil.as_str() {
-                    "$" => SymbolKind::scalar(),
-                    "@" => SymbolKind::array(),
-                    "%" => SymbolKind::hash(),
-                    _ => return None,
-                };
-
-                let symbols = self.symbol_table.find_symbol(name, 0, kind);
-                symbols.first()?;
-
-                // Return the basic type based on sigil
-                match sigil.as_str() {
-                    "$" => Some("scalar".to_string()),
-                    "@" => Some("array".to_string()),
-                    "%" => Some("hash".to_string()),
-                    _ => None,
-                }
-            }
-
-            NodeKind::FunctionCall { name, .. } => {
-                // Basic return type inference for built-in functions
-                match name.as_str() {
-                    "scalar" => Some("scalar".to_string()),
-                    "ref" => Some("string".to_string()),
-                    "length" | "index" | "rindex" => Some("number".to_string()),
-                    "split" => Some("array".to_string()),
-                    "keys" | "values" => Some("array".to_string()),
-                    _ => None,
-                }
-            }
-
-            NodeKind::Binary { op, .. } => {
-                // Infer based on operator
-                match op.as_str() {
-                    "+" | "-" | "*" | "/" | "%" | "**" => Some("number".to_string()),
-                    "." | "x" => Some("string".to_string()),
-                    "==" | "!=" | "<" | ">" | "<=" | ">=" | "eq" | "ne" | "lt" | "gt" | "le"
-                    | "ge" => Some("boolean".to_string()),
-                    _ => None,
-                }
-            }
-
-            _ => None,
-        }
-    }
-}
-
-/// Documentation entry for a Perl built-in function.
-///
-/// Provides signature and description information for display in hover tooltips.
-pub struct BuiltinDoc {
-    /// Function signature showing calling conventions
-    pub signature: &'static str,
-    /// Brief description of what the function does
-    pub description: &'static str,
-}
-
-/// Check if a function name is a Perl built-in.
-///
-/// Returns `true` if the name matches a known Perl built-in function.
-fn is_control_keyword(name: &str) -> bool {
-    matches!(name, "next" | "last" | "redo" | "goto" | "return" | "exit" | "die")
-}
-
-fn is_builtin_function(name: &str) -> bool {
-    matches!(
-        name,
-        "print"
-            | "say"
-            | "printf"
-            | "sprintf"
-            | "open"
-            | "close"
-            | "read"
-            | "write"
-            | "chomp"
-            | "chop"
-            | "split"
-            | "join"
-            | "push"
-            | "pop"
-            | "shift"
-            | "unshift"
-            | "sort"
-            | "reverse"
-            | "map"
-            | "grep"
-            | "length"
-            | "substr"
-            | "index"
-            | "rindex"
-            | "lc"
-            | "uc"
-            | "lcfirst"
-            | "ucfirst"
-            | "defined"
-            | "undef"
-            | "ref"
-            | "blessed"
-            | "die"
-            | "warn"
-            | "eval"
-            | "require"
-            | "use"
-            | "return"
-            | "next"
-            | "last"
-            | "redo"
-            | "goto" // ... many more
-    )
-}
-
-/// Get documentation for a Perl built-in function.
-///
-/// Returns signature and description for known built-in functions,
-/// or `None` if documentation is not available.
-///
-/// This is also used by the LSP hover handler to show builtin docs when the
-/// semantic analyzer has no symbol-level hit (e.g. bare-word builtins in
-/// fallback path).
-pub fn get_builtin_documentation(name: &str) -> Option<BuiltinDoc> {
-    match name {
-        // I/O
-        "print" => Some(BuiltinDoc {
-            signature: "print FILEHANDLE LIST\nprint LIST\nprint",
-            description: "Prints a string or list of strings. If FILEHANDLE is omitted, prints to the last selected output handle (STDOUT by default).",
-        }),
-        "say" => Some(BuiltinDoc {
-            signature: "say FILEHANDLE LIST\nsay LIST\nsay",
-            description: "Like print, but appends a newline to the output.",
-        }),
-        "printf" => Some(BuiltinDoc {
-            signature: "printf FILEHANDLE FORMAT, LIST\nprintf FORMAT, LIST",
-            description: "Prints a formatted string to FILEHANDLE (default STDOUT).",
-        }),
-        "sprintf" => Some(BuiltinDoc {
-            signature: "sprintf FORMAT, LIST",
-            description: "Returns a formatted string (like C sprintf). Does not print.",
-        }),
-        "open" => Some(BuiltinDoc {
-            signature: "open FILEHANDLE, MODE, EXPR\nopen FILEHANDLE, EXPR\nopen FILEHANDLE",
-            description: "Opens the file whose filename is given by EXPR, and associates it with FILEHANDLE.",
-        }),
-        "close" => Some(BuiltinDoc {
-            signature: "close FILEHANDLE\nclose",
-            description: "Closes the file, socket, or pipe associated with FILEHANDLE.",
-        }),
-        "read" => Some(BuiltinDoc {
-            signature: "read FILEHANDLE, SCALAR, LENGTH, OFFSET\nread FILEHANDLE, SCALAR, LENGTH",
-            description: "Reads LENGTH bytes of data into SCALAR from FILEHANDLE. Returns the number of bytes read, or undef on error.",
-        }),
-        "write" => Some(BuiltinDoc {
-            signature: "write FILEHANDLE\nwrite",
-            description: "Writes a formatted record to FILEHANDLE using the format associated with it.",
-        }),
-        "seek" => Some(BuiltinDoc {
-            signature: "seek FILEHANDLE, POSITION, WHENCE",
-            description: "Sets the position for a filehandle. WHENCE: 0=start, 1=current, 2=end.",
-        }),
-        "tell" => Some(BuiltinDoc {
-            signature: "tell FILEHANDLE\ntell",
-            description: "Returns the current position in bytes for FILEHANDLE.",
-        }),
-        "eof" => Some(BuiltinDoc {
-            signature: "eof FILEHANDLE\neof()\neof",
-            description: "Returns true if the next read on FILEHANDLE would return end of file.",
-        }),
-        "binmode" => Some(BuiltinDoc {
-            signature: "binmode FILEHANDLE, LAYER\nbinmode FILEHANDLE",
-            description: "Sets binary mode on FILEHANDLE, or specifies an I/O layer.",
-        }),
-        "truncate" => Some(BuiltinDoc {
-            signature: "truncate FILEHANDLE, LENGTH",
-            description: "Truncates the file at the given LENGTH.",
-        }),
-
-        // String functions
-        "chomp" => Some(BuiltinDoc {
-            signature: "chomp VARIABLE\nchomp LIST\nchomp",
-            description: "Removes the trailing newline from VARIABLE. Returns the number of characters removed.",
-        }),
-        "chop" => Some(BuiltinDoc {
-            signature: "chop VARIABLE\nchop LIST\nchop",
-            description: "Removes and returns the last character from VARIABLE.",
-        }),
-        "length" => Some(BuiltinDoc {
-            signature: "length EXPR\nlength",
-            description: "Returns the length in characters of the value of EXPR.",
-        }),
-        "substr" => Some(BuiltinDoc {
-            signature: "substr EXPR, OFFSET, LENGTH, REPLACEMENT\nsubstr EXPR, OFFSET, LENGTH\nsubstr EXPR, OFFSET",
-            description: "Extracts a substring out of EXPR and returns it. With REPLACEMENT, replaces the substring in-place.",
-        }),
-        "index" => Some(BuiltinDoc {
-            signature: "index STR, SUBSTR, POSITION\nindex STR, SUBSTR",
-            description: "Returns the position of the first occurrence of SUBSTR in STR at or after POSITION. Returns -1 if not found.",
-        }),
-        "rindex" => Some(BuiltinDoc {
-            signature: "rindex STR, SUBSTR, POSITION\nrindex STR, SUBSTR",
-            description: "Returns the position of the last occurrence of SUBSTR in STR at or before POSITION.",
-        }),
-        "lc" => Some(BuiltinDoc {
-            signature: "lc EXPR\nlc",
-            description: "Returns a lowercased version of EXPR (or $_ if omitted).",
-        }),
-        "uc" => Some(BuiltinDoc {
-            signature: "uc EXPR\nuc",
-            description: "Returns an uppercased version of EXPR (or $_ if omitted).",
-        }),
-        "lcfirst" => Some(BuiltinDoc {
-            signature: "lcfirst EXPR\nlcfirst",
-            description: "Returns EXPR with the first character lowercased.",
-        }),
-        "ucfirst" => Some(BuiltinDoc {
-            signature: "ucfirst EXPR\nucfirst",
-            description: "Returns EXPR with the first character uppercased.",
-        }),
-        "chr" => Some(BuiltinDoc {
-            signature: "chr NUMBER\nchr",
-            description: "Returns the character represented by NUMBER in the character set.",
-        }),
-        "ord" => Some(BuiltinDoc {
-            signature: "ord EXPR\nord",
-            description: "Returns the numeric value of the first character of EXPR.",
-        }),
-        "hex" => Some(BuiltinDoc {
-            signature: "hex EXPR\nhex",
-            description: "Interprets EXPR as a hex string and returns the corresponding numeric value.",
-        }),
-        "oct" => Some(BuiltinDoc {
-            signature: "oct EXPR\noct",
-            description: "Interprets EXPR as an octal string and returns the corresponding value. Handles 0x, 0b, and 0 prefixes.",
-        }),
-        "quotemeta" => Some(BuiltinDoc {
-            signature: "quotemeta EXPR\nquotemeta",
-            description: "Returns EXPR with all non-alphanumeric characters backslashed (escaped for regex).",
-        }),
-        "join" => Some(BuiltinDoc {
-            signature: "join EXPR, LIST",
-            description: "Joins the separate strings of LIST into a single string with fields separated by EXPR, and returns that string.",
-        }),
-        "split" => Some(BuiltinDoc {
-            signature: "split /PATTERN/, EXPR, LIMIT\nsplit /PATTERN/, EXPR\nsplit /PATTERN/\nsplit",
-            description: "Splits the string EXPR into a list of strings and returns the list. If LIMIT is specified, splits into at most that many fields.",
-        }),
-
-        // Array/List
-        "push" => Some(BuiltinDoc {
-            signature: "push ARRAY, LIST",
-            description: "Appends one or more values to the end of ARRAY. Returns the number of elements in the resulting array.",
-        }),
-        "pop" => Some(BuiltinDoc {
-            signature: "pop ARRAY\npop",
-            description: "Removes and returns the last element of ARRAY.",
-        }),
-        "shift" => Some(BuiltinDoc {
-            signature: "shift ARRAY\nshift",
-            description: "Removes and returns the first element of ARRAY, shortening the array by 1.",
-        }),
-        "unshift" => Some(BuiltinDoc {
-            signature: "unshift ARRAY, LIST",
-            description: "Prepends LIST to the front of ARRAY. Returns the number of elements in the new array.",
-        }),
-        "splice" => Some(BuiltinDoc {
-            signature: "splice ARRAY, OFFSET, LENGTH, LIST\nsplice ARRAY, OFFSET, LENGTH\nsplice ARRAY, OFFSET\nsplice ARRAY",
-            description: "Removes LENGTH elements from ARRAY starting at OFFSET, replacing them with LIST. Returns the removed elements.",
-        }),
-        "sort" => Some(BuiltinDoc {
-            signature: "sort SUBNAME LIST\nsort BLOCK LIST\nsort LIST",
-            description: "Sorts LIST and returns the sorted list. BLOCK or SUBNAME provides a custom comparison function using $a and $b.",
-        }),
-        "reverse" => Some(BuiltinDoc {
-            signature: "reverse LIST",
-            description: "In list context, returns LIST in reverse order. In scalar context, returns a string with characters reversed.",
-        }),
-        "map" => Some(BuiltinDoc {
-            signature: "map BLOCK LIST\nmap EXPR, LIST",
-            description: "Evaluates the BLOCK or EXPR for each element of LIST (locally setting $_ to each element) and composes a list of the results.",
-        }),
-        "grep" => Some(BuiltinDoc {
-            signature: "grep BLOCK LIST\ngrep EXPR, LIST",
-            description: "Evaluates BLOCK or EXPR for each element of LIST and returns the list of elements for which the expression is true.",
-        }),
-        "scalar" => Some(BuiltinDoc {
-            signature: "scalar EXPR",
-            description: "Forces EXPR to be interpreted in scalar context and returns the value of EXPR.",
-        }),
-        "wantarray" => Some(BuiltinDoc {
-            signature: "wantarray",
-            description: "Returns true if the currently executing subroutine is expected to return a list value.",
-        }),
-
-        // Hash
-        "keys" => Some(BuiltinDoc {
-            signature: "keys HASH\nkeys ARRAY",
-            description: "Returns a list of all the keys of the named hash, or the indices of an array.",
-        }),
-        "values" => Some(BuiltinDoc {
-            signature: "values HASH\nvalues ARRAY",
-            description: "Returns a list of all the values of the named hash, or values of an array.",
-        }),
-        "each" => Some(BuiltinDoc {
-            signature: "each HASH\neach ARRAY",
-            description: "Returns a two-element list of the next (key, value) pair from the hash.",
-        }),
-        "exists" => Some(BuiltinDoc {
-            signature: "exists EXPR",
-            description: "Returns true if the specified hash key or array element exists, even if its value is undef.",
-        }),
-        "delete" => Some(BuiltinDoc {
-            signature: "delete EXPR",
-            description: "Deletes the specified keys and their associated values from a hash, or elements from an array.",
-        }),
-        "defined" => Some(BuiltinDoc {
-            signature: "defined EXPR\ndefined",
-            description: "Returns true if EXPR has a value other than undef.",
-        }),
-        "undef" => Some(BuiltinDoc {
-            signature: "undef EXPR\nundef",
-            description: "Undefines the value of EXPR. Can be used on scalars, arrays, hashes, subroutines, and typeglobs.",
-        }),
-
-        // References and OO
-        "ref" => Some(BuiltinDoc {
-            signature: "ref EXPR\nref",
-            description: "Returns a string indicating the type of reference EXPR is, or empty string if not a reference. E.g. HASH, ARRAY, SCALAR, CODE.",
-        }),
-        "bless" => Some(BuiltinDoc {
-            signature: "bless REF, CLASSNAME\nbless REF",
-            description: "Associates the referent of REF with package CLASSNAME (or current package). Returns the reference.",
-        }),
-        "blessed" => Some(BuiltinDoc {
-            signature: "blessed EXPR",
-            description: "Returns the name of the package EXPR is blessed into, or undef if EXPR is not a blessed reference. From Scalar::Util.",
-        }),
-        "tie" => Some(BuiltinDoc {
-            signature: "tie VARIABLE, CLASSNAME, LIST",
-            description: "Binds a variable to a package class that provides the implementation for the variable.",
-        }),
-        "untie" => Some(BuiltinDoc {
-            signature: "untie VARIABLE",
-            description: "Breaks the binding between a variable and its package.",
-        }),
-        "tied" => Some(BuiltinDoc {
-            signature: "tied VARIABLE",
-            description: "Returns a reference to the object underlying VARIABLE if it is tied, or undef if not.",
-        }),
-
-        // Control flow
-        "die" => Some(BuiltinDoc {
-            signature: "die LIST",
-            description: "Raises an exception. If the last element of LIST does not end in a newline, the current script line number and input line number are appended.",
-        }),
-        "warn" => Some(BuiltinDoc {
-            signature: "warn LIST",
-            description: "Produces a message on STDERR, like die but does not exit or throw an exception.",
-        }),
-        "eval" => Some(BuiltinDoc {
-            signature: "eval BLOCK\neval EXPR",
-            description: "Evaluates EXPR or BLOCK and traps any errors, making them available in $@.",
-        }),
-        "return" => Some(BuiltinDoc {
-            signature: "return EXPR\nreturn",
-            description: "Returns from a subroutine with the value of EXPR.",
-        }),
-        "next" => Some(BuiltinDoc {
-            signature: "next LABEL\nnext",
-            description: "Starts the next iteration of the loop (like C 'continue').",
-        }),
-        "last" => Some(BuiltinDoc {
-            signature: "last LABEL\nlast",
-            description: "Exits the loop immediately (like C 'break').",
-        }),
-        "redo" => Some(BuiltinDoc {
-            signature: "redo LABEL\nredo",
-            description: "Restarts the loop block without re-evaluating the condition.",
-        }),
-        "goto" => Some(BuiltinDoc {
-            signature: "goto LABEL\ngoto EXPR\ngoto &NAME",
-            description: "Transfers control to the named label, computed label, or substitutes a call to the named subroutine.",
-        }),
-        "caller" => Some(BuiltinDoc {
-            signature: "caller EXPR\ncaller",
-            description: "Returns information about the calling subroutine's context. Returns (package, filename, line) in list context.",
-        }),
-        "exit" => Some(BuiltinDoc {
-            signature: "exit EXPR\nexit",
-            description: "Exits the program with status EXPR (default 0). Calls END blocks and DESTROY methods before exit.",
-        }),
-
-        // Modules and loading
-        "require" => Some(BuiltinDoc {
-            signature: "require EXPR\nrequire",
-            description: "Loads a library module at runtime. Raises an exception on failure.",
-        }),
-        "use" => Some(BuiltinDoc {
-            signature: "use Module VERSION LIST\nuse Module VERSION\nuse Module LIST\nuse Module",
-            description: "Loads and imports a module at compile time. Equivalent to BEGIN { require Module; Module->import( LIST ); }",
-        }),
-        "do" => Some(BuiltinDoc {
-            signature: "do BLOCK\ndo EXPR",
-            description: "As do BLOCK: executes BLOCK and returns its value. As do EXPR: reads and executes a Perl file.",
-        }),
-
-        // Math
-        "abs" => Some(BuiltinDoc {
-            signature: "abs VALUE\nabs",
-            description: "Returns the absolute value of its argument.",
-        }),
-        "int" => Some(BuiltinDoc {
-            signature: "int EXPR\nint",
-            description: "Returns the integer portion of EXPR (truncates toward zero).",
-        }),
-        "sqrt" => Some(BuiltinDoc {
-            signature: "sqrt EXPR\nsqrt",
-            description: "Returns the positive square root of EXPR.",
-        }),
-        "log" => Some(BuiltinDoc {
-            signature: "log EXPR\nlog",
-            description: "Returns the natural logarithm (base e) of EXPR.",
-        }),
-        "exp" => Some(BuiltinDoc {
-            signature: "exp EXPR\nexp",
-            description: "Returns e (the natural logarithm base) to the power of EXPR.",
-        }),
-        "sin" => Some(BuiltinDoc {
-            signature: "sin EXPR\nsin",
-            description: "Returns the sine of EXPR (expressed in radians).",
-        }),
-        "cos" => Some(BuiltinDoc {
-            signature: "cos EXPR\ncos",
-            description: "Returns the cosine of EXPR (expressed in radians).",
-        }),
-        "atan2" => Some(BuiltinDoc {
-            signature: "atan2 Y, X",
-            description: "Returns the arctangent of Y/X in the range -PI to PI.",
-        }),
-        "rand" => Some(BuiltinDoc {
-            signature: "rand EXPR\nrand",
-            description: "Returns a random fractional number greater than or equal to 0 and less than EXPR (default 1).",
-        }),
-        "srand" => Some(BuiltinDoc {
-            signature: "srand EXPR\nsrand",
-            description: "Sets the random number seed for the rand operator.",
-        }),
-
-        // File tests and operations
-        "stat" => Some(BuiltinDoc {
-            signature: "stat FILEHANDLE\nstat EXPR",
-            description: "Returns a 13-element list giving the status info for a file. (dev, ino, mode, nlink, uid, gid, rdev, size, atime, mtime, ctime, blksize, blocks).",
-        }),
-        "lstat" => Some(BuiltinDoc {
-            signature: "lstat FILEHANDLE\nlstat EXPR",
-            description: "Like stat, but if the last component of the filename is a symbolic link, stats the link itself.",
-        }),
-        "chmod" => Some(BuiltinDoc {
-            signature: "chmod MODE, LIST",
-            description: "Changes the permissions of a list of files. Returns the number of files successfully changed.",
-        }),
-        "chown" => Some(BuiltinDoc {
-            signature: "chown UID, GID, LIST",
-            description: "Changes the owner and group of a list of files.",
-        }),
-        "unlink" => Some(BuiltinDoc {
-            signature: "unlink LIST\nunlink",
-            description: "Deletes a list of files. Returns the number of files successfully deleted.",
-        }),
-        "rename" => Some(BuiltinDoc {
-            signature: "rename OLDNAME, NEWNAME",
-            description: "Renames a file. Returns true on success, false otherwise.",
-        }),
-        "mkdir" => Some(BuiltinDoc {
-            signature: "mkdir FILENAME, MODE\nmkdir FILENAME",
-            description: "Creates the directory specified by FILENAME. Returns true on success.",
-        }),
-        "rmdir" => Some(BuiltinDoc {
-            signature: "rmdir FILENAME\nrmdir",
-            description: "Deletes the directory if it is empty. Returns true on success.",
-        }),
-        "opendir" => Some(BuiltinDoc {
-            signature: "opendir DIRHANDLE, EXPR",
-            description: "Opens a directory for reading by readdir.",
-        }),
-        "readdir" => Some(BuiltinDoc {
-            signature: "readdir DIRHANDLE",
-            description: "Returns the next entry (or entries in list context) from the directory.",
-        }),
-        "closedir" => Some(BuiltinDoc {
-            signature: "closedir DIRHANDLE",
-            description: "Closes a directory opened by opendir.",
-        }),
-        "link" => Some(BuiltinDoc {
-            signature: "link OLDFILE, NEWFILE",
-            description: "Creates a new hard link for an existing file.",
-        }),
-        "symlink" => Some(BuiltinDoc {
-            signature: "symlink OLDFILE, NEWFILE",
-            description: "Creates a new symbolic link for an existing file.",
-        }),
-        "readlink" => Some(BuiltinDoc {
-            signature: "readlink EXPR\nreadlink",
-            description: "Returns the value of a symbolic link.",
-        }),
-        "chdir" => Some(BuiltinDoc {
-            signature: "chdir EXPR\nchdir",
-            description: "Changes the working directory to EXPR (or home directory if omitted).",
-        }),
-        "glob" => Some(BuiltinDoc {
-            signature: "glob EXPR\nglob",
-            description: "Returns the filenames matching the shell-style glob pattern EXPR.",
-        }),
-
-        // System/Process
-        "system" => Some(BuiltinDoc {
-            signature: "system LIST\nsystem PROGRAM LIST",
-            description: "Executes a system command and returns the exit status. The return value is the exit status of the program as returned by the wait call.",
-        }),
-        "exec" => Some(BuiltinDoc {
-            signature: "exec LIST\nexec PROGRAM LIST",
-            description: "Replaces the current process with an external command. Never returns on success.",
-        }),
-        "fork" => Some(BuiltinDoc {
-            signature: "fork",
-            description: "Creates a child process. Returns the child pid to the parent, 0 to the child, or undef on failure.",
-        }),
-        "wait" => Some(BuiltinDoc {
-            signature: "wait",
-            description: "Waits for a child process to terminate and returns the pid of the deceased process.",
-        }),
-        "waitpid" => Some(BuiltinDoc {
-            signature: "waitpid PID, FLAGS",
-            description: "Waits for a particular child process to terminate and returns the pid.",
-        }),
-        "kill" => Some(BuiltinDoc {
-            signature: "kill SIGNAL, LIST",
-            description: "Sends a signal to a list of processes. Returns the number of processes signalled.",
-        }),
-        "sleep" => Some(BuiltinDoc {
-            signature: "sleep EXPR\nsleep",
-            description: "Causes the script to sleep for EXPR seconds (or forever if no argument).",
-        }),
-        "alarm" => Some(BuiltinDoc {
-            signature: "alarm SECONDS\nalarm",
-            description: "Arranges to have a SIGALRM delivered after SECONDS seconds.",
-        }),
-
-        // Encoding/Decoding
-        "pack" => Some(BuiltinDoc {
-            signature: "pack TEMPLATE, LIST",
-            description: "Takes a list of values and packs it into a binary string according to TEMPLATE.",
-        }),
-        "unpack" => Some(BuiltinDoc {
-            signature: "unpack TEMPLATE, EXPR",
-            description: "Takes a binary string and expands it into a list of values according to TEMPLATE.",
-        }),
-        "crypt" => Some(BuiltinDoc {
-            signature: "crypt PLAINTEXT, SALT",
-            description: "Encrypts a string using the system crypt() function.",
-        }),
-
-        // Time
-        "time" => Some(BuiltinDoc {
-            signature: "time",
-            description: "Returns the number of seconds since the epoch (January 1, 1970 UTC).",
-        }),
-        "localtime" => Some(BuiltinDoc {
-            signature: "localtime EXPR\nlocaltime",
-            description: "Converts a time value to a 9-element list with the time analyzed for the local time zone. In scalar context returns a ctime(3) string.",
-        }),
-        "gmtime" => Some(BuiltinDoc {
-            signature: "gmtime EXPR\ngmtime",
-            description: "Like localtime but uses Greenwich Mean Time (UTC).",
-        }),
-
-        // Misc
-        "prototype" => Some(BuiltinDoc {
-            signature: "prototype FUNCTION",
-            description: "Returns the prototype of a function as a string, or undef if the function has no prototype.",
-        }),
-        "local" => Some(BuiltinDoc {
-            signature: "local EXPR",
-            description: "Temporarily localizes the listed global variables to the enclosing block. The original values are restored at the end of the block.",
-        }),
-        "my" => Some(BuiltinDoc {
-            signature: "my VARLIST\nmy TYPE VARLIST",
-            description: "Declares lexically scoped variables. Variables are visible only within the enclosing block.",
-        }),
-        "our" => Some(BuiltinDoc {
-            signature: "our VARLIST",
-            description: "Declares package variables visible in the current lexical scope without qualifying the name.",
-        }),
-        "state" => Some(BuiltinDoc {
-            signature: "state VARLIST",
-            description: "Declares lexically scoped variables that persist across calls to the enclosing subroutine (like C static variables).",
-        }),
-
-        _ => None,
-    }
-}
-
-/// Documentation entry for a Perl subroutine or variable attribute.
-///
-/// Provides a human-readable description of what an attribute does, used to
-/// enrich hover tooltips in the IDE with semantic meaning beyond the raw
-/// attribute name.
-pub struct PerlAttributeDoc {
-    /// The canonical attribute name (without leading colon)
-    pub name: &'static str,
-    /// Human-readable description of what the attribute does
-    pub description: &'static str,
-}
-
-/// Get documentation for a Perl subroutine or variable attribute.
-///
-/// Accepts the attribute name with or without a leading colon (`:lvalue` or `lvalue`).
-/// Returns `None` for unknown or custom attributes.
-///
-/// # Built-in Attributes Covered
-/// - `:lvalue` — allows subroutines to be used as lvalues (assigned to)
-/// - `:method` — marks a subroutine as an object method
-/// - `:prototype` — overrides the subroutine's calling prototype
-/// - `:const` — makes a subroutine a compile-time constant
-/// - `:shared` — marks a variable as shared between threads (`threads::shared`)
-/// - `:locked` — acquires a mutex on the variable before invoking the subroutine (deprecated)
-/// - `:unique` — marks a variable as unique across the interpreter (deprecated, removed Perl 5.28)
-pub fn get_attribute_documentation(attr: &str) -> Option<PerlAttributeDoc> {
-    // Strip a leading colon so callers can pass either ":lvalue" or "lvalue"
-    let name = attr.strip_prefix(':').unwrap_or(attr);
-    match name {
-        "lvalue" => Some(PerlAttributeDoc {
-            name: "lvalue",
-            description: "Marks the subroutine as an lvalue: it can appear on the left-hand side \
-                          of an assignment. The subroutine should return a value that can be \
-                          assigned to.",
-        }),
-        "method" => Some(PerlAttributeDoc {
-            name: "method",
-            description: "Marks the subroutine as an object method. Used by some call-sequence \
-                          optimisers and documentation tools; does not change runtime dispatch.",
-        }),
-        "prototype" => Some(PerlAttributeDoc {
-            name: "prototype",
-            description: "Overrides the subroutine's calling prototype. The argument supplies the \
-                          prototype string, e.g. :prototype($$) for two scalar arguments.",
-        }),
-        "const" => Some(PerlAttributeDoc {
-            name: "const",
-            description: "Makes the subroutine a compile-time constant folder. When called with a \
-                          constant argument, the result is inlined at compile time. Implies \
-                          read-only semantics.",
-        }),
-        "shared" => Some(PerlAttributeDoc {
-            name: "shared",
-            description: "Marks the variable as shared between threads (requires threads::shared). \
-                          Access is automatically synchronised across all threads.",
-        }),
-        "locked" => Some(PerlAttributeDoc {
-            name: "locked",
-            description: "Deprecated. Previously caused a mutex to be acquired on the subroutine \
-                          or variable before it was invoked/accessed. Has no effect in modern Perl.",
-        }),
-        "unique" => Some(PerlAttributeDoc {
-            name: "unique",
-            description: "Deprecated. Previously marked a variable as unique across the \
-                          interpreter. Has no effect in modern Perl (removed in Perl 5.28).",
-        }),
-        _ => None,
-    }
-}
-
-/// Build enriched hover detail lines for a list of Perl attributes.
-///
-/// For each attribute in the list this function looks up the semantic
-/// documentation via [`get_attribute_documentation`].  If documentation
-/// is found, the line is formatted as `":name — description"`.  If no
-/// documentation is found (custom/third-party attribute) the raw attribute
-/// name is included unchanged.
-///
-/// Returns an empty `Vec` when `attributes` is empty.
-pub(crate) fn format_attribute_details(attributes: &[String]) -> Vec<String> {
-    if attributes.is_empty() {
-        return Vec::new();
-    }
-    let mut result = Vec::with_capacity(attributes.len() + 1);
-    result.push("Attributes:".to_string());
-    for attr in attributes {
-        let canonical = attr.strip_prefix(':').unwrap_or(attr.as_str());
-        match get_attribute_documentation(canonical) {
-            Some(doc) => result.push(format!("{} — {}", attr, doc.description)),
-            None => result.push(format!("Attribute: {}", attr)),
-        }
-    }
-    result
-}
-
-#[derive(Debug)]
-/// A stable, query-oriented view of semantic information over a parsed file.
-///
-/// LSP and other consumers should use this instead of talking to `SemanticAnalyzer` directly.
-/// This provides a clean API that insulates consumers from internal analyzer implementation details.
-///
-/// # Performance Characteristics
-/// - Symbol resolution: <50μs average lookup time
-/// - Reference queries: O(1) lookup via pre-computed indices
-/// - Scope queries: O(log n) with binary search on scope ranges
-///
-/// # LSP Workflow Integration
-/// Core component in Parse → Index → Navigate → Complete → Analyze pipeline:
-/// 1. Parse Perl source → AST
-/// 2. Build SemanticModel from AST
-/// 3. Query for symbols, references, completions
-/// 4. Respond to LSP requests with precise semantic data
-///
-/// # Example
-/// ```rust,ignore
-/// use perl_parser::Parser;
-/// use perl_parser::semantic::SemanticModel;
-///
-/// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-/// let code = "my $x = 42; $x + 10;";
-/// let mut parser = Parser::new(code);
-/// let ast = parser.parse()?;
-///
-/// let model = SemanticModel::build(&ast, code);
-/// let tokens = model.tokens();
-/// assert!(!tokens.is_empty());
-/// # Ok(())
-/// # }
-/// ```
-pub struct SemanticModel {
-    /// Internal semantic analyzer instance
-    analyzer: SemanticAnalyzer,
-}
-
-impl SemanticModel {
-    /// Build a semantic model for a parsed syntax tree.
-    ///
-    /// # Parameters
-    /// - `root`: The root AST node from the parser
-    /// - `source`: The original Perl source code
-    ///
-    /// # Performance
-    /// - Analysis time: O(n) where n is AST node count
-    /// - Memory: ~1MB per 10K lines of Perl code
-    pub fn build(root: &Node, source: &str) -> Self {
-        Self { analyzer: SemanticAnalyzer::analyze_with_source(root, source) }
-    }
-
-    /// All semantic tokens for syntax highlighting.
-    ///
-    /// Returns tokens in source order for efficient LSP semantic tokens encoding.
-    ///
-    /// # Performance
-    /// - Lookup: O(1) - pre-computed during analysis
-    /// - Memory: ~32 bytes per token
-    pub fn tokens(&self) -> &[SemanticToken] {
-        self.analyzer.semantic_tokens()
-    }
-
-    /// Access the underlying symbol table for advanced queries.
-    ///
-    /// # Note
-    /// Most consumers should use the higher-level query methods on `SemanticModel`
-    /// rather than accessing the symbol table directly.
-    pub fn symbol_table(&self) -> &SymbolTable {
-        self.analyzer.symbol_table()
-    }
-
-    /// Get hover information for a symbol at a specific location during Navigate/Analyze.
-    ///
-    /// # Parameters
-    /// - `location`: Source location to query (line, column)
-    ///
-    /// # Returns
-    /// - `Some(HoverInfo)` if a symbol with hover info exists at this location
-    /// - `None` if no symbol or no hover info available
-    ///
-    /// # Performance
-    /// - Lookup: <100μs for typical files
-    /// - Memory: Cached hover info reused across queries
-    ///
-    /// Workflow: Navigate/Analyze hover lookup.
-    pub fn hover_info_at(&self, location: SourceLocation) -> Option<&HoverInfo> {
-        self.analyzer.hover_at(location)
-    }
-
-    /// Find the definition of a symbol at a specific byte position.
-    ///
-    /// # Parameters
-    /// - `position`: Byte offset in the source code
-    ///
-    /// # Returns
-    /// - `Some(Symbol)` if a symbol definition is found at this position
-    /// - `None` if no symbol exists at this position
-    ///
-    /// # Performance
-    /// - Lookup: <50μs average for typical files
-    /// - Uses pre-computed symbol table for O(1) lookups
-    ///
-    /// # Example
-    /// ```rust,ignore
-    /// use perl_parser::Parser;
-    /// use perl_parser::semantic::SemanticModel;
-    ///
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// let code = "my $x = 1;\n$x + 2;\n";
-    /// let mut parser = Parser::new(code);
-    /// let ast = parser.parse()?;
-    ///
-    /// let model = SemanticModel::build(&ast, code);
-    /// // Find definition of $x on line 1 (byte position ~11)
-    /// if let Some(symbol) = model.definition_at(11) {
-    ///     assert_eq!(symbol.location.start.line, 0);
-    /// }
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn definition_at(&self, position: usize) -> Option<&Symbol> {
-        self.analyzer.find_definition(position)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::parser::Parser;
-
-    #[test]
-    fn test_semantic_tokens() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-my $x = 42;
-print $x;
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-
-        let analyzer = SemanticAnalyzer::analyze(&ast);
-        let tokens = analyzer.semantic_tokens();
-
-        // Phase 1 implementation (Issue #188) handles critical AST node types
-        // including VariableListDeclaration, Ternary, ArrayLiteral, HashLiteral,
-        // Try, and PhaseBlock nodes
-
-        // Check first $x is a declaration
-        let x_tokens: Vec<_> = tokens
-            .iter()
-            .filter(|t| {
-                matches!(
-                    t.token_type,
-                    SemanticTokenType::Variable | SemanticTokenType::VariableDeclaration
-                )
-            })
-            .collect();
-        assert!(!x_tokens.is_empty());
-        assert!(x_tokens[0].modifiers.contains(&SemanticTokenModifier::Declaration));
-        Ok(())
-    }
-
-    #[test]
-    fn test_hover_info() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-sub foo {
-    return 42;
-}
-
-my $result = foo();
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-
-        let analyzer = SemanticAnalyzer::analyze(&ast);
-
-        // The hover info would be at specific locations
-        // In practice, we'd look up by position
-        assert!(!analyzer.hover_info.is_empty());
-        Ok(())
-    }
-
-    #[test]
-    fn test_hover_doc_from_pod() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-# This is foo
-# More docs
-sub foo {
-    return 1;
-}
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-
-        let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-
-        // Find the symbol for foo and check its hover documentation
-        let sym = analyzer.symbol_table().symbols.get("foo").ok_or("symbol not found")?[0].clone();
-        let hover = analyzer.hover_at(sym.location).ok_or("hover not found")?;
-        assert!(hover.documentation.as_ref().ok_or("doc not found")?.contains("This is foo"));
-        Ok(())
-    }
-
-    #[test]
-    fn test_comment_doc_extraction() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-# Adds two numbers
-sub add { 1 }
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-
-        let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-
-        let sub_symbols =
-            analyzer.symbol_table().find_symbol("add", 0, crate::symbol::SymbolKind::Subroutine);
-        assert!(!sub_symbols.is_empty());
-        let hover = analyzer.hover_at(sub_symbols[0].location).ok_or("hover not found")?;
-        assert_eq!(hover.documentation.as_deref(), Some("Adds two numbers"));
-        Ok(())
-    }
-
-    #[test]
-    fn test_cross_package_navigation() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-package Foo {
-    # bar sub
-    sub bar { 42 }
-}
-
-package main;
-Foo::bar();
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-        let pos = code.find("Foo::bar").ok_or("Foo::bar not found")? + 5; // position within "bar"
-        let def = analyzer.find_definition(pos).ok_or("definition")?;
-        assert_eq!(def.name, "bar");
-
-        let hover = analyzer.hover_at(def.location).ok_or("hover not found")?;
-        assert!(hover.documentation.as_ref().ok_or("doc not found")?.contains("bar sub"));
-        Ok(())
-    }
-
-    #[test]
-    fn test_scope_identification() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-my $x = 0;
-package Foo {
-    my $x = 1;
-    sub bar { return $x; }
-}
-my $y = $x;
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-
-        let inner_ref_pos = code.find("return $x").ok_or("return $x not found")? + "return ".len();
-        let inner_def = analyzer.find_definition(inner_ref_pos).ok_or("inner def not found")?;
-        let expected_inner = code.find("my $x = 1").ok_or("my $x = 1 not found")? + 3;
-        assert_eq!(inner_def.location.start, expected_inner);
-
-        let outer_ref_pos = code.rfind("$x;").ok_or("$x; not found")?;
-        let outer_def = analyzer.find_definition(outer_ref_pos).ok_or("outer def not found")?;
-        let expected_outer = code.find("my $x = 0").ok_or("my $x = 0 not found")? + 3;
-        assert_eq!(outer_def.location.start, expected_outer);
-        Ok(())
-    }
-
-    #[test]
-    fn test_pod_documentation_extraction() -> Result<(), Box<dyn std::error::Error>> {
-        // Test with a simple case that parses correctly
-        let code = r#"# Simple comment before sub
-sub documented_with_comment {
-    return "test";
-}
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-
-        let sub_symbols = analyzer.symbol_table().find_symbol(
-            "documented_with_comment",
-            0,
-            crate::symbol::SymbolKind::Subroutine,
-        );
-        assert!(!sub_symbols.is_empty());
-        let hover = analyzer.hover_at(sub_symbols[0].location).ok_or("hover not found")?;
-        let doc = hover.documentation.as_ref().ok_or("doc not found")?;
-        assert!(doc.contains("Simple comment before sub"));
-        Ok(())
-    }
-
-    #[test]
-    fn test_empty_source_handling() -> Result<(), Box<dyn std::error::Error>> {
-        let code = "";
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-
-        // Should not crash with empty source
-        assert!(analyzer.semantic_tokens().is_empty());
-        assert!(analyzer.hover_info.is_empty());
-        Ok(())
-    }
-
-    #[test]
-    fn test_multiple_comment_lines() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-# First comment
-# Second comment
-# Third comment
-sub multi_commented {
-    1;
-}
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-
-        let sub_symbols = analyzer.symbol_table().find_symbol(
-            "multi_commented",
-            0,
-            crate::symbol::SymbolKind::Subroutine,
-        );
-        assert!(!sub_symbols.is_empty());
-        let hover = analyzer.hover_at(sub_symbols[0].location).ok_or("hover not found")?;
-        let doc = hover.documentation.as_ref().ok_or("doc not found")?;
-        assert!(doc.contains("First comment"));
-        assert!(doc.contains("Second comment"));
-        assert!(doc.contains("Third comment"));
-        Ok(())
-    }
-
-    // SemanticModel tests
-    #[test]
-    fn test_semantic_model_build_and_tokens() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-my $x = 42;
-my $y = 10;
-$x + $y;
-"#;
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-
-        let model = SemanticModel::build(&ast, code);
-
-        // Should have semantic tokens
-        let tokens = model.tokens();
-        assert!(!tokens.is_empty(), "SemanticModel should provide tokens");
-
-        // Should have variable tokens
-        let var_tokens: Vec<_> = tokens
-            .iter()
-            .filter(|t| {
-                matches!(
-                    t.token_type,
-                    SemanticTokenType::Variable | SemanticTokenType::VariableDeclaration
-                )
-            })
-            .collect();
-        assert!(var_tokens.len() >= 2, "Should have at least 2 variable tokens");
-        Ok(())
-    }
-
-    #[test]
-    fn test_semantic_model_symbol_table_access() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-my $x = 42;
-sub foo {
-    my $y = $x;
-}
-"#;
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-
-        let model = SemanticModel::build(&ast, code);
-
-        // Should be able to access symbol table
-        let symbol_table = model.symbol_table();
-        let x_symbols = symbol_table.find_symbol("x", 0, SymbolKind::scalar());
-        assert!(!x_symbols.is_empty(), "Should find $x in symbol table");
-
-        let foo_symbols = symbol_table.find_symbol("foo", 0, SymbolKind::Subroutine);
-        assert!(!foo_symbols.is_empty(), "Should find sub foo in symbol table");
-        Ok(())
-    }
-
-    #[test]
-    fn test_semantic_model_hover_info() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-# This is a documented variable
-my $documented = 42;
-"#;
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-
-        let model = SemanticModel::build(&ast, code);
-
-        // Find the location of the variable declaration
-        let symbol_table = model.symbol_table();
-        let symbols = symbol_table.find_symbol("documented", 0, SymbolKind::scalar());
-        assert!(!symbols.is_empty(), "Should find $documented");
-
-        // Check if hover info is available
-        if let Some(hover) = model.hover_info_at(symbols[0].location) {
-            assert!(hover.signature.contains("documented"), "Hover should contain variable name");
-        }
-        // Note: hover_info_at might return None if no explicit hover was generated,
-        // which is acceptable for now
-        Ok(())
-    }
-
-    #[test]
-    fn test_analyzer_find_definition_scalar() -> Result<(), Box<dyn std::error::Error>> {
-        let code = "my $x = 1;\n$x + 2;\n";
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-
-        // Use the same path SemanticModel uses to feed source
-        let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-
-        // Find the byte offset of the reference "$x" in the second line
-        let ref_line = code.lines().nth(1).ok_or("line 2 not found")?;
-        let line_offset = code.lines().next().ok_or("line 1 not found")?.len() + 1; // +1 for '\n'
-        let col_in_line = ref_line.find("$x").ok_or("could not find $x on line 2")?;
-        let ref_pos = line_offset + col_in_line;
-
-        let symbol =
-            analyzer.find_definition(ref_pos).ok_or("definition not found for $x reference")?;
-
-        // 1. Must be a scalar named "x"
-        assert_eq!(symbol.name, "x");
-        assert_eq!(symbol.kind, SymbolKind::scalar());
-
-        // 2. Declaration must come before reference
-        assert!(
-            symbol.location.start < ref_pos,
-            "Declaration {:?} should precede reference at byte {}",
-            symbol.location.start,
-            ref_pos
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_semantic_model_definition_at() -> Result<(), Box<dyn std::error::Error>> {
-        let code = "my $x = 1;\n$x + 2;\n";
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-
-        let model = SemanticModel::build(&ast, code);
-
-        // Compute the byte offset of the reference "$x" on the second line
-        let ref_line_index = 1;
-        let ref_line = code.lines().nth(ref_line_index).ok_or("line not found")?;
-        let col_in_line = ref_line.find("$x").ok_or("could not find $x")?;
-        let byte_offset = code
-            .lines()
-            .take(ref_line_index)
-            .map(|l| l.len() + 1) // +1 for '\n'
-            .sum::<usize>()
-            + col_in_line;
-
-        let definition = model.definition_at(byte_offset);
-        assert!(
-            definition.is_some(),
-            "definition_at returned None for $x reference at {}",
-            byte_offset
-        );
-        if let Some(symbol) = definition {
-            assert_eq!(symbol.name, "x");
-            assert_eq!(symbol.kind, SymbolKind::scalar());
-            assert!(
-                symbol.location.start < byte_offset,
-                "Declaration {:?} should precede reference at byte {}",
-                symbol.location.start,
-                byte_offset
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_anonymous_subroutine_semantic_tokens() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-my $closure = sub {
-    my $x = 42;
-    return $x + 1;
-};
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-
-        // Check that we have semantic tokens for the anonymous sub
-        let tokens = analyzer.semantic_tokens();
-
-        // Should have a keyword token for 'sub'
-        let sub_keywords: Vec<_> =
-            tokens.iter().filter(|t| matches!(t.token_type, SemanticTokenType::Keyword)).collect();
-
-        assert!(!sub_keywords.is_empty(), "Should have keyword token for 'sub'");
-
-        // Check hover info exists for the anonymous sub
-        let sub_position = code.find("sub {").ok_or("sub { not found")?;
-        let hover_exists = analyzer
-            .hover_info
-            .iter()
-            .any(|(loc, _)| loc.start <= sub_position && loc.end >= sub_position);
-
-        assert!(hover_exists, "Should have hover info for anonymous subroutine");
-        Ok(())
-    }
-
-    #[test]
-    fn test_infer_type_for_literals() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-my $num = 42;
-my $str = "hello";
-my @arr = (1, 2, 3);
-my %hash = (a => 1);
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-
-        // Find nodes and test type inference
-        // We need to walk the AST to find the literal nodes
-        fn find_number_node(node: &Node) -> Option<&Node> {
-            match &node.kind {
-                NodeKind::Number { .. } => Some(node),
-                NodeKind::Program { statements } | NodeKind::Block { statements } => {
-                    for stmt in statements {
-                        if let Some(found) = find_number_node(stmt) {
-                            return Some(found);
-                        }
-                    }
-                    None
-                }
-                NodeKind::VariableDeclaration { initializer, .. } => {
-                    initializer.as_ref().and_then(|init| find_number_node(init))
-                }
-                _ => None,
-            }
-        }
-
-        if let Some(num_node) = find_number_node(&ast) {
-            let inferred = analyzer.infer_type(num_node);
-            assert_eq!(inferred, Some("number".to_string()), "Should infer number type");
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_infer_type_for_binary_operations() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"my $sum = 10 + 20;
-my $concat = "a" . "b";
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-
-        // Find binary operation nodes
-        fn find_binary_node<'a>(node: &'a Node, op: &str) -> Option<&'a Node> {
-            match &node.kind {
-                NodeKind::Binary { op: node_op, .. } if node_op == op => Some(node),
-                NodeKind::Program { statements } | NodeKind::Block { statements } => {
-                    for stmt in statements {
-                        if let Some(found) = find_binary_node(stmt, op) {
-                            return Some(found);
-                        }
-                    }
-                    None
-                }
-                NodeKind::VariableDeclaration { initializer, .. } => {
-                    initializer.as_ref().and_then(|init| find_binary_node(init, op))
-                }
-                _ => None,
-            }
-        }
-
-        // Test arithmetic operation infers to number
-        if let Some(add_node) = find_binary_node(&ast, "+") {
-            let inferred = analyzer.infer_type(add_node);
-            assert_eq!(inferred, Some("number".to_string()), "Arithmetic should infer to number");
-        }
-
-        // Test concatenation infers to string
-        if let Some(concat_node) = find_binary_node(&ast, ".") {
-            let inferred = analyzer.infer_type(concat_node);
-            assert_eq!(
-                inferred,
-                Some("string".to_string()),
-                "Concatenation should infer to string"
-            );
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_anonymous_subroutine_hover_info() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-# This is a closure
-my $adder = sub {
-    my ($x, $y) = @_;
-    return $x + $y;
-};
-"#;
-
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-
-        // Find hover info for the anonymous sub
-        let sub_position = code.find("sub {").ok_or("sub { not found")?;
-        let hover = analyzer
-            .hover_info
-            .iter()
-            .find(|(loc, _)| loc.start <= sub_position && loc.end >= sub_position)
-            .map(|(_, h)| h);
-
-        assert!(hover.is_some(), "Should have hover info");
-
-        if let Some(h) = hover {
-            assert!(h.signature.contains("sub"), "Hover signature should contain 'sub'");
-            assert!(
-                h.details.iter().any(|d| d.contains("Anonymous")),
-                "Hover details should mention anonymous subroutine"
-            );
-            // Documentation extraction searches backwards from the sub keyword,
-            // but the comment is before `my $adder =` (not immediately before `sub`),
-            // so extract_documentation may not find it. Accept either outcome.
-            if let Some(doc) = &h.documentation {
-                assert!(
-                    doc.contains("closure"),
-                    "If documentation found, it should mention closure"
-                );
-            }
-        }
-        Ok(())
-    }
-
-    // Phase 2/3 Handler Tests
-    #[test]
-    fn test_substitution_operator_semantic_token() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-my $str = "hello world";
-$str =~ s/world/Perl/;
-"#;
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze(&ast);
-
-        let tokens = analyzer.semantic_tokens();
-        let operator_tokens: Vec<_> =
-            tokens.iter().filter(|t| matches!(t.token_type, SemanticTokenType::Operator)).collect();
-
-        assert!(!operator_tokens.is_empty(), "Should have operator tokens for substitution");
-        Ok(())
-    }
-
-    #[test]
-    fn test_transliteration_operator_semantic_token() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-my $str = "hello";
-$str =~ tr/el/ol/;
-"#;
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze(&ast);
-
-        let tokens = analyzer.semantic_tokens();
-        let operator_tokens: Vec<_> =
-            tokens.iter().filter(|t| matches!(t.token_type, SemanticTokenType::Operator)).collect();
-
-        assert!(!operator_tokens.is_empty(), "Should have operator tokens for transliteration");
-        Ok(())
-    }
-
-    #[test]
-    fn test_reference_operator_semantic_token() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-my $x = 42;
-my $ref = \$x;
-"#;
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze(&ast);
-
-        let tokens = analyzer.semantic_tokens();
-        let operator_tokens: Vec<_> =
-            tokens.iter().filter(|t| matches!(t.token_type, SemanticTokenType::Operator)).collect();
-
-        assert!(!operator_tokens.is_empty(), "Should have operator tokens for reference operator");
-        Ok(())
-    }
-
-    #[test]
-    fn test_postfix_loop_semantic_token() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-my @list = (1, 2, 3);
-print $_ for @list;
-my $x = 0;
-$x++ while $x < 10;
-"#;
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze(&ast);
-
-        let tokens = analyzer.semantic_tokens();
-        let control_tokens: Vec<_> = tokens
-            .iter()
-            .filter(|t| matches!(t.token_type, SemanticTokenType::KeywordControl))
-            .collect();
-
-        assert!(!control_tokens.is_empty(), "Should have control keyword tokens for postfix loops");
-        Ok(())
-    }
-
-    #[test]
-    fn test_file_test_operator_semantic_token() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-my $file = "test.txt";
-if (-e $file) {
-    print "exists";
-}
-if (-d $file) {
-    print "directory";
-}
-if (-f $file) {
-    print "file";
-}
-"#;
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze(&ast);
-
-        let tokens = analyzer.semantic_tokens();
-        let operator_tokens: Vec<_> =
-            tokens.iter().filter(|t| matches!(t.token_type, SemanticTokenType::Operator)).collect();
-
-        assert!(!operator_tokens.is_empty(), "Should have operator tokens for file test operators");
-        Ok(())
-    }
-
-    #[test]
-    fn test_all_file_test_operators_recognized() -> Result<(), Box<dyn std::error::Error>> {
-        // Test that the is_file_test_operator helper recognizes all file test operators
-        let file_test_ops = vec![
-            "-e", "-d", "-f", "-r", "-w", "-x", "-s", "-z", "-T", "-B", "-M", "-A", "-C", "-l",
-            "-p", "-S", "-u", "-g", "-k", "-t", "-O", "-G", "-R", "-b", "-c",
-        ];
-
-        for op in file_test_ops {
-            assert!(
-                SemanticAnalyzer::is_file_test_operator(op),
-                "Operator {} should be recognized as file test operator",
-                op
-            );
-        }
-
-        // Test that non-file-test operators are not recognized
-        assert!(
-            !SemanticAnalyzer::is_file_test_operator("+"),
-            "Operator '+' should not be recognized as file test operator"
-        );
-        assert!(
-            !SemanticAnalyzer::is_file_test_operator("-"),
-            "Operator '-' should not be recognized as file test operator"
-        );
-        assert!(
-            !SemanticAnalyzer::is_file_test_operator("++"),
-            "Operator '++' should not be recognized as file test operator"
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_postfix_loop_modifiers() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-my @items = (1, 2, 3);
-print $_ for @items;
-print $_ foreach @items;
-my $x = 0;
-$x++ while $x < 10;
-$x-- until $x < 0;
-"#;
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze(&ast);
-
-        let tokens = analyzer.semantic_tokens();
-        let control_tokens: Vec<_> = tokens
-            .iter()
-            .filter(|t| matches!(t.token_type, SemanticTokenType::KeywordControl))
-            .collect();
-
-        // Should have at least 4 control keyword tokens (for, foreach, while, until)
-        assert!(
-            control_tokens.len() >= 4,
-            "Should have at least 4 control keyword tokens for postfix loop modifiers"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_substitution_with_modifiers() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-my $str = "hello world";
-$str =~ s/world/Perl/gi;
-"#;
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze(&ast);
-
-        let tokens = analyzer.semantic_tokens();
-        let operator_tokens: Vec<_> =
-            tokens.iter().filter(|t| matches!(t.token_type, SemanticTokenType::Operator)).collect();
-
-        assert!(
-            !operator_tokens.is_empty(),
-            "Should have operator tokens for substitution with modifiers"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_transliteration_y_operator() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-my $str = "hello";
-$str =~ y/hello/world/;
-"#;
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze(&ast);
-
-        let tokens = analyzer.semantic_tokens();
-        let operator_tokens: Vec<_> =
-            tokens.iter().filter(|t| matches!(t.token_type, SemanticTokenType::Operator)).collect();
-
-        assert!(
-            !operator_tokens.is_empty(),
-            "Should have operator tokens for y/// transliteration"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_builtin_documentation_coverage() -> Result<(), Box<dyn std::error::Error>> {
-        // Verify that commonly-used builtins have documentation
-        let builtins = [
-            "print", "say", "push", "pop", "shift", "unshift", "map", "grep", "sort", "reverse",
-            "split", "join", "chomp", "chop", "length", "substr", "index", "rindex", "lc", "uc",
-            "die", "warn", "eval", "open", "close", "read", "keys", "values", "exists", "delete",
-            "defined", "ref", "bless", "sprintf", "chr", "ord",
-        ];
-
-        for name in &builtins {
-            let doc = get_builtin_documentation(name);
-            assert!(doc.is_some(), "Built-in '{}' should have documentation", name);
-            let doc = doc.unwrap();
-            assert!(
-                !doc.signature.is_empty(),
-                "Built-in '{}' should have a non-empty signature",
-                name
-            );
-            assert!(
-                !doc.description.is_empty(),
-                "Built-in '{}' should have a non-empty description",
-                name
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_builtin_hover_for_function_call() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-my @items = (3, 1, 4);
-push @items, 5;
-"#;
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-
-        // Find the hover info for 'push' function call
-        let push_pos = code.find("push").ok_or("push not found")?;
-        let hover_for_push =
-            analyzer.hover_info.iter().find(|(loc, _)| loc.start <= push_pos && loc.end > push_pos);
-
-        assert!(hover_for_push.is_some(), "Should have hover info for 'push' builtin");
-        let (_, hover) = hover_for_push.unwrap();
-        assert!(
-            hover.signature.contains("push"),
-            "Hover signature should contain 'push', got: {}",
-            hover.signature
-        );
-        assert!(hover.documentation.is_some(), "Hover for 'push' should have documentation");
-        Ok(())
-    }
-
-    #[test]
-    fn test_package_hover_with_pod_name_section() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-=head1 NAME
-
-My::Module - A great module for testing
-
-=head1 DESCRIPTION
-
-This module does great things.
-
-=cut
-
-package My::Module;
-
-sub new { bless {}, shift }
-
-1;
-"#;
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-
-        // Find the package symbol
-        let pkg_symbols = analyzer.symbol_table().symbols.get("My::Module");
-        assert!(pkg_symbols.is_some(), "Should find My::Module in symbol table");
-
-        let pkg = &pkg_symbols.unwrap()[0];
-        let hover = analyzer.hover_at(pkg.location);
-        assert!(hover.is_some(), "Should have hover info for package");
-
-        let hover = hover.unwrap();
-        assert!(
-            hover.signature.contains("package My::Module"),
-            "Package hover signature should contain 'package My::Module', got: {}",
-            hover.signature
-        );
-        // The POD NAME section should be extracted as documentation
-        if let Some(doc) = &hover.documentation {
-            assert!(
-                doc.contains("A great module for testing"),
-                "Package hover should contain POD NAME content, got: {}",
-                doc
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_package_documentation_via_symbol() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-=head1 NAME
-
-Utils - Utility functions
-
-=cut
-
-package Utils;
-
-sub helper { 1 }
-
-1;
-"#;
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-
-        let pkg_symbols = analyzer.symbol_table().symbols.get("Utils");
-        assert!(pkg_symbols.is_some(), "Should find Utils package");
-
-        let pkg = &pkg_symbols.unwrap()[0];
-        // The symbol extractor should have extracted POD docs
-        assert!(
-            pkg.documentation.is_some(),
-            "Package symbol should have documentation from POD NAME section"
-        );
-        let doc = pkg.documentation.as_ref().unwrap();
-        assert!(
-            doc.contains("Utility functions"),
-            "Package doc should contain 'Utility functions', got: {}",
-            doc
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_subroutine_with_pod_docs_hover() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"
-=head2 process
-
-Processes input data and returns the result.
-
-=cut
-
-sub process {
-    my ($input) = @_;
-    return $input * 2;
-}
-"#;
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-
-        let sub_symbols = analyzer.symbol_table().find_symbol("process", 0, SymbolKind::Subroutine);
-        assert!(!sub_symbols.is_empty(), "Should find sub process");
-
-        let hover = analyzer.hover_at(sub_symbols[0].location);
-        assert!(hover.is_some(), "Should have hover for sub process");
-
-        let hover = hover.unwrap();
-        assert!(
-            hover.signature.contains("sub process"),
-            "Hover should show sub signature, got: {}",
-            hover.signature
-        );
-        // The POD =head2 docs should be extracted
-        if let Some(doc) = &hover.documentation {
-            assert!(
-                doc.contains("process") || doc.contains("Processes"),
-                "Sub hover should contain POD documentation, got: {}",
-                doc
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_variable_hover_shows_declaration_type() -> Result<(), Box<dyn std::error::Error>> {
-        let code = r#"my $count = 42;
-my @items = (1, 2, 3);
-my %config = (key => "value");
-"#;
-        let mut parser = Parser::new(code);
-        let ast = parser.parse()?;
-        let analyzer = SemanticAnalyzer::analyze_with_source(&ast, code);
-
-        // Check scalar variable hover
-        let scalar_pos = code.find("$count").ok_or("$count not found")?;
-        let scalar_hover = analyzer
-            .hover_info
-            .iter()
-            .find(|(loc, _)| loc.start <= scalar_pos && loc.end > scalar_pos);
-        assert!(scalar_hover.is_some(), "Should have hover for $count");
-        let (_, hover) = scalar_hover.unwrap();
-        assert!(
-            hover.signature.contains("$count"),
-            "Scalar hover should show variable name, got: {}",
-            hover.signature
-        );
-
-        // Check array variable hover
-        let array_pos = code.find("@items").ok_or("@items not found")?;
-        let array_hover = analyzer
-            .hover_info
-            .iter()
-            .find(|(loc, _)| loc.start <= array_pos && loc.end > array_pos);
-        assert!(array_hover.is_some(), "Should have hover for @items");
-        let (_, hover) = array_hover.unwrap();
-        assert!(
-            hover.signature.contains("@items"),
-            "Array hover should show variable name, got: {}",
-            hover.signature
-        );
-
-        // Check hash variable hover
-        let hash_pos = code.find("%config").ok_or("%config not found")?;
-        let hash_hover =
-            analyzer.hover_info.iter().find(|(loc, _)| loc.start <= hash_pos && loc.end > hash_pos);
-        assert!(hash_hover.is_some(), "Should have hover for %config");
-        let (_, hover) = hash_hover.unwrap();
-        assert!(
-            hover.signature.contains("%config"),
-            "Hash hover should show variable name, got: {}",
-            hover.signature
-        );
-        Ok(())
-    }
-}
